+++ paq8pxd.cpp
@@ -26,7 +26,7 @@
 #ifndef DISABLE_SM
 //#define SM              // For faster statemap
 #endif
-#define VERBOSE         // Show extra info
+//#define VERBOSE         // Show extra info
 
 #ifdef WINDOWS                       
 #ifdef MT
@@ -128,6 +128,9 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "prt/contextmap2.hpp"
 #include "prt/contextmap.hpp"
 #include "prt/bh.hpp"
+//#include "prt/wrt/wrton.hpp"
+#include "prt/stemmer/stemmer.hpp"
+#include "prt/tables.hpp"
 
 #include "models/model.hpp"
 #include "models/sparsematch.hpp"
@@ -154,6 +157,7 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "models/im24bit.hpp"
 #include "models/blank.hpp"
 #include "models/lstm.hpp"
+#include "models/word.hpp"
 
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
@@ -246,27 +250,7 @@ const U8 WRT_mtt[256]= {
  7,7,7,7,7,7,7,7, 7,7,7,7,7,7,7,7,
  7,7,7,7,7,7,7,7, 7,7,7,7,7,7,7,7,
 };
-const U8 WRT_wrd1[256]={
-   0,  0,  0, 64,  0, 64,192,  0,   0,  0, 64,  0,  0,  0,  0,  0,
-   0,  0,  0,  0,  0,  0,  0,  0, 192,  0,  0,  0,  0,  0,  0,  0,
-   0,  0,192,  0,  0,  0,  0,  0,   0,  0,  0,  0,128,  0,  0,  0,
-  64, 64, 64, 64, 64, 64, 64, 64,  64, 64,  0,  0,128,  0,128,  0,
 
-   0, 64, 64, 64, 64, 64, 64, 64,  64, 64, 64, 64, 64, 64, 64, 64,
-  64, 64, 64, 64, 64, 64, 64, 64,  64, 64, 64,  0,192,  0,  0,  0,
-   0, 64, 64, 64, 64, 64, 64, 64,  64, 64, 64, 64, 64, 64, 64, 64,
-  64, 64, 64, 64, 64, 64, 64, 64,  64, 64, 64,  0,  0,  0,  0,  0,
-
- 128,128,128,128,128,128,128,128, 128,128,128,128,128,128,128,128,
- 128,128,128,128,128,128,128,128, 128,128,128,128,128,128,128,128,
- 128,128,128,128,128,128,128,128, 128,128,128,128,128,128,128,128,
- 128,128,128,128,128,128,128,128, 128,128,128,128,128,128,128,128,
- 128,128,128,128,128,128,128,128, 128,128,128,128,128,128,128,128,
-
- 192,192,192,192,192,192,192,192, 192,192,192,192,192,192,192,192,
- 192,192,192,192,192,192,192,192, 192,192,192,192,192,192,192,192,
- 192,192,192,192,192,192,192,192, 192,192,192,192,192,192,192,192,
- };
 const U32 tri[4]={0,4,3,7}, trj[4]={0,6,6,12};
 
 int dt[1024];  // i -> 16K/(i+i+3)
@@ -281,1467 +265,9 @@ int n0n1[256]; // for contectmap
 
 //////////////////////////// Text modelling /////////////////////////
 
-inline bool CharInArray(const char c, const char a[], const int len) {
-  if (*a==0)
-    return false;
-  int i=0;
-  for (; i<len && c!=a[i]; i++);
-  return i<len;
-}
-
-#define MAX_WORD_SIZE 64
 
 
-class Word {
-public:
-  U8 Letters[MAX_WORD_SIZE];
-  U8 Start, End;
-  U32 Hash[4], Type, Language;
-  Word() : Start(0), End(0),  Type(0), Language(0) {
-    memset(&Letters[0], 0, sizeof(U8)*MAX_WORD_SIZE);
-    memset(&Hash[0], 0, sizeof(U32)*4);
-  }
-  bool operator==(const char *s) const {
-    size_t len=strlen(s);
-    return ((size_t)(End-Start+(Letters[Start]!=0))==len && memcmp(&Letters[Start], s, len)==0);
-  }
-  bool operator!=(const char *s) const {
-    return !operator==(s);
-  }
-  void operator+=(const char c) {
-    if (End<MAX_WORD_SIZE-1){
-      End+=(Letters[End]>0);
-      Letters[End]=tolower(c);
-    }
-  }
-  U8 operator[](U8 i) const {
-    return (End-Start>=i)?Letters[Start+i]:0;
-  }
-  U8 operator()(U8 i) const {
-    return (End-Start>=i)?Letters[End-i]:0;
-  }
-  U32 Length() const {
-    if (Letters[Start]!=0)
-      return End-Start+1;
-    return 0;
-  }
-  void GetHashes() {
-    Hash[0] = 0xc01dflu, Hash[1] = ~Hash[0];
-    for (int i=Start; i<=End; i++) {
-      U8 l = Letters[i];
-      Hash[0]^=hash(Hash[0], l, i);
-      Hash[1]^=hash(Hash[1], 
-        ((l&0x80)==0)?l&0x5F:
-        ((l&0xC0)==0x80)?l&0x3F:
-        ((l&0xE0)==0xC0)?l&0x1F:
-        ((l&0xF0)==0xE0)?l&0xF:l&0x7
-      );
-    }
-    Hash[2] = (~Hash[0])^Hash[1];
-    Hash[3] = (~Hash[1])^Hash[0];
-  }
-  bool ChangeSuffix(const char *OldSuffix, const char *NewSuffix) {
-    size_t len=strlen(OldSuffix);
-    if (Length()>len && memcmp(&Letters[End-len+1], OldSuffix, len)==0){
-      size_t n=strlen(NewSuffix);
-      if (n>0){
-        memcpy(&Letters[End-int(len)+1], NewSuffix, min(MAX_WORD_SIZE-1,End+int(n))-End);
-        End=min(MAX_WORD_SIZE-1, End-int(len)+int(n));
-      }
-      else
-        End-=U8(len);
-      return true;
-    }
-    return false;
-  }
-  bool MatchesAny(const char* a[], const int count) {
-    int i=0;
-    size_t len = (size_t)Length();
-    for (; i<count && (len!=strlen(a[i]) || memcmp(&Letters[Start], a[i], len)!=0); i++);
-    return i<count;
-  }
-  bool EndsWith(const char *Suffix) const {
-    size_t len=strlen(Suffix);
-    return (Length()>len && memcmp(&Letters[End-len+1], Suffix, len)==0);
-  }
-  bool StartsWith(const char *Prefix) const {
-    size_t len=strlen(Prefix);
-    return (Length()>len && memcmp(&Letters[Start], Prefix, len)==0);
-  }
-  void print() const {
-    for(int r=Start;r<=End;r++)
-      printf("%c",(char)Letters[r]);
-    printf("\n");
-  }
-};
 
-class Segment1 {
-public:
-  Word FirstWord; // useful following questions
-  U32 WordCount;
-  U32 NumCount;
-};
-
-class Sentence : public Segment1 {
-public:
-  enum Types { // possible sentence types, excluding Imperative
-    Declarative,
-    Interrogative,
-    Exclamative,
-    Count
-  };
-  Types Type;
-  U32 SegmentCount;
-  U32 VerbIndex; // relative position of last detected verb
-  U32 NounIndex; // relative position of last detected noun
-  U32 CapitalIndex; // relative position of last capitalized word, excluding the initial word of this sentence
-  Word lastVerb, lastNoun, lastCapital;
-};
-
-class Paragraph {
-public:
-  U32 SentenceCount, TypeCount[Sentence::Types::Count], TypeMask;
-};
-
-class Language {
-public:
-   enum Flags {
-    Verb                   = (1<<0),
-    Noun                   = (1<<1)
-  };
-  enum Ids {
-    Unknown,
-    English,
-    French,
-    German,
-    Count
-  };
-  virtual ~Language() {};
-  virtual bool IsAbbreviation(Word *W) = 0;
-};
-
-class English: public Language {
-private:
-  static const int NUM_ABBREV = 6;
-  const char *Abbreviations[NUM_ABBREV]={ "mr","mrs","ms","dr","st","jr" };
-public:
-  enum Flags {
-   
-    Adjective              = (1<<2),
-    Plural                 = (1<<3),
-    Male                   = (1<<4),
-    Female                 = (1<<5),
-    Negation               = (1<<6),
-    PastTense              = (1<<7)|Verb,
-    PresentParticiple      = (1<<8)|Verb,
-    AdjectiveSuperlative   = (1<<9)|Adjective,
-    AdjectiveWithout       = (1<<10)|Adjective,
-    AdjectiveFull          = (1<<11)|Adjective,
-    AdverbOfManner         = (1<<12),
-    SuffixNESS             = (1<<13),
-    SuffixITY              = (1<<14)|Noun,
-    SuffixCapable          = (1<<15),
-    SuffixNCE              = (1<<16),
-    SuffixNT               = (1<<17),
-    SuffixION              = (1<<18),
-    SuffixAL               = (1<<19)|Adjective,
-    SuffixIC               = (1<<20)|Adjective,
-    SuffixIVE              = (1<<21),
-    SuffixOUS              = (1<<22)|Adjective,
-    PrefixOver             = (1<<23),
-    PrefixUnder            = (1<<24)
-  };
-  bool IsAbbreviation(Word *W) { return W->MatchesAny(Abbreviations, NUM_ABBREV); };
-};
-
-class French: public Language {
-private:
-  static const int NUM_ABBREV = 2;
-  const char *Abbreviations[NUM_ABBREV]={ "m","mm" };
-public:
-  enum Flags {
-    
-    Adjective              = (1<<2),
-    Plural                 = (1<<3)
-  };
-  bool IsAbbreviation(Word *W) { return W->MatchesAny(Abbreviations, NUM_ABBREV); };
-};
-
-class German : public Language {
-private:
-  static const int NUM_ABBREV = 3;
-  const char *Abbreviations[NUM_ABBREV]={ "fr","hr","hrn" };
-public:
-  enum Flags {
-    Adjective              = (1<<2),
-    Plural                 = (1<<3),
-    Female                 = (1<<4)
-  };
-  bool IsAbbreviation(Word *W) { return W->MatchesAny(Abbreviations, NUM_ABBREV); };
-};
-//////////////////////////// Stemming routines /////////////////////////
-
-class Stemmer {
-    protected:
-  U32 GetRegion(const Word *W, const U32 From) {
-    bool hasVowel = false;
-    for (int i=W->Start+From; i<=W->End; i++) {
-      if (IsVowel(W->Letters[i])) {
-        hasVowel = true;
-        continue;
-      }
-      else if (hasVowel)
-        return i-W->Start+1;
-    }
-    return W->Start+W->Length();
-  }
-  bool SuffixInRn(const Word *W, const U32 Rn, const char *Suffix) {
-    return (W->Start!=W->End && Rn<=W->Length()-strlen(Suffix));
-  }
-public:
-  virtual ~Stemmer() {};
-  virtual bool IsVowel(const char c) = 0;
-  virtual void Hash(Word *W) = 0;
-  virtual bool Stem(Word *W) = 0;
-};
-
-
-class EnglishStemmer: public Stemmer {
-private:
-  static const int NUM_VOWELS = 6;
-  const char Vowels[NUM_VOWELS]={'a','e','i','o','u','y'};
-  static const int NUM_DOUBLES = 9;
-  const char Doubles[NUM_DOUBLES]={'b','d','f','g','m','n','p','r','t'};
-  static const int NUM_LI_ENDINGS = 10;
-  const char LiEndings[NUM_LI_ENDINGS]={'c','d','e','g','h','k','m','n','r','t'};
-  static const int NUM_NON_SHORT_CONSONANTS = 3;
-  const char NonShortConsonants[NUM_NON_SHORT_CONSONANTS]={'w','x','Y'};
-  static const int NUM_MALE_WORDS = 9;
-  const char *MaleWords[NUM_MALE_WORDS]={"he","him","his","himself","man","men","boy","husband","actor"};
-  static const int NUM_FEMALE_WORDS = 8;
-  const char *FemaleWords[NUM_FEMALE_WORDS]={"she","her","herself","woman","women","girl","wife","actress"};
-  static const int NUM_COMMON_WORDS = 12;
-  const char *CommonWords[NUM_COMMON_WORDS]={"the","be","to","of","and","in","that","you","have","with","from","but"};
-  static const int NUM_SUFFIXES_STEP0 = 3;
-  const char *SuffixesStep0[NUM_SUFFIXES_STEP0]={"'s'","'s","'"};
-  static const int NUM_SUFFIXES_STEP1b = 6;
-  const char *SuffixesStep1b[NUM_SUFFIXES_STEP1b]={"eedly","eed","ed","edly","ing","ingly"};
-  const U32 TypesStep1b[NUM_SUFFIXES_STEP1b]={English::AdverbOfManner,0,English::PastTense,English::AdverbOfManner|English::PastTense,English::PresentParticiple,English::AdverbOfManner|English::PresentParticiple};
-  static const int NUM_SUFFIXES_STEP2 = 22;
-  const char *(SuffixesStep2[NUM_SUFFIXES_STEP2])[2]={
-    {"ization", "ize"},
-    {"ational", "ate"},
-    {"ousness", "ous"},
-    {"iveness", "ive"},
-    {"fulness", "ful"},
-    {"tional", "tion"},
-    {"lessli", "less"},
-    {"biliti", "ble"},
-    {"entli", "ent"},
-    {"ation", "ate"},
-    {"alism", "al"},
-    {"aliti", "al"},
-    {"fulli", "ful"},
-    {"ousli", "ous"},
-    {"iviti", "ive"},
-    {"enci", "ence"},
-    {"anci", "ance"},
-    {"abli", "able"},
-    {"izer", "ize"},
-    {"ator", "ate"},
-    {"alli", "al"},
-    {"bli", "ble"}
-  };
-  const U32 TypesStep2[NUM_SUFFIXES_STEP2]={
-    English::SuffixION,
-    English::SuffixION|English::SuffixAL,
-    English::SuffixNESS,
-    English::SuffixNESS,
-    English::SuffixNESS,
-    English::SuffixION|English::SuffixAL,
-    English::AdverbOfManner,
-    English::AdverbOfManner|English::SuffixITY,
-    English::AdverbOfManner,
-    English::SuffixION,
-    0,
-    English::SuffixITY,
-    English::AdverbOfManner,
-    English::AdverbOfManner,
-    English::SuffixITY,
-    0,
-    0,
-    English::AdverbOfManner,
-    0,
-    0,
-    English::AdverbOfManner,
-    English::AdverbOfManner
-  };
-  static const int NUM_SUFFIXES_STEP3 = 8;
-  const char *(SuffixesStep3[NUM_SUFFIXES_STEP3])[2]={
-    {"ational", "ate"},
-    {"tional", "tion"},
-    {"alize", "al"},
-    {"icate", "ic"},
-    {"iciti", "ic"},
-    {"ical", "ic"},
-    {"ful", ""},
-    {"ness", ""}
-  };
-  const U32 TypesStep3[NUM_SUFFIXES_STEP3]={English::SuffixION|English::SuffixAL,English::SuffixION|English::SuffixAL,0,0,English::SuffixITY,English::SuffixAL,English::AdjectiveFull,English::SuffixNESS};
-  static const int NUM_SUFFIXES_STEP4 = 20;
-  const char *SuffixesStep4[NUM_SUFFIXES_STEP4]={"al","ance","ence","er","ic","able","ible","ant","ement","ment","ent","ou","ism","ate","iti","ous","ive","ize","sion","tion"};
-  const U32 TypesStep4[NUM_SUFFIXES_STEP4]={
-    English::SuffixAL,
-    English::SuffixNCE,
-    English::SuffixNCE,
-    0,
-    English::SuffixIC,
-    English::SuffixCapable,
-    English::SuffixCapable,
-    English::SuffixNT,
-    0,
-    0,
-    English::SuffixNT,
-    0,
-    0,
-    0,
-    English::SuffixITY,
-    English::SuffixOUS,
-    English::SuffixIVE,
-    0,
-    English::SuffixION,
-    English::SuffixION
-  };
-  static const int NUM_EXCEPTION_REGION1 = 3;
-  const char *ExceptionsRegion1[NUM_EXCEPTION_REGION1]={"gener","arsen","commun"};
-  static const int NUM_EXCEPTIONS1 = 19;
-  const char *(Exceptions1[NUM_EXCEPTIONS1])[2]={
-    {"skis", "ski"},
-    {"skies", "sky"},
-    {"dying", "die"},
-    {"lying", "lie"},
-    {"tying", "tie"},
-    {"idly", "idl"},
-    {"gently", "gentl"},
-    {"ugly", "ugli"},
-    {"early", "earli"},
-    {"only", "onli"},
-    {"singly", "singl"},
-    {"sky", "sky"},
-    {"news", "news"},
-    {"howe", "howe"},
-    {"atlas", "atlas"},
-    {"cosmos", "cosmos"},
-    {"bias", "bias"},
-    {"andes", "andes"},
-    {"texas", "texas"}
-  };
-  const U32 TypesExceptions1[NUM_EXCEPTIONS1]={
-    English::Noun|English::Plural,
-    English::Noun|English::Plural,
-    English::PresentParticiple,
-    English::PresentParticiple,
-    English::PresentParticiple,
-    English::AdverbOfManner,
-    English::AdverbOfManner,
-    English::Adjective,
-    English::Adjective|English::AdverbOfManner,
-    0,
-    English::AdverbOfManner,
-    English::Noun,
-    English::Noun,
-    0,
-    English::Noun,
-    English::Noun,
-    English::Noun,
-    English::Noun|English::Plural,
-    English::Noun
-  };
-  static const int NUM_EXCEPTIONS2 = 8;
-  const char *Exceptions2[NUM_EXCEPTIONS2]={"inning","outing","canning","herring","earring","proceed","exceed","succeed"};
-  const U32 TypesExceptions2[NUM_EXCEPTIONS2]={English::Noun,English::Noun,English::Noun,English::Noun,English::Noun,English::Verb,English::Verb,English::Verb}; 
-  inline bool IsConsonant(const char c){
-    return !IsVowel(c);
-  }
-  inline bool IsShortConsonant(const char c){
-    return !CharInArray(c, NonShortConsonants, NUM_NON_SHORT_CONSONANTS);
-  }
-  inline bool IsDouble(const char c){
-    return CharInArray(c, Doubles, NUM_DOUBLES);
-  }
-  inline bool IsLiEnding(const char c){
-    return CharInArray(c, LiEndings, NUM_LI_ENDINGS);
-  }
-   
-  U32 GetRegion1(const Word *W){
-    for (int i=0; i<NUM_EXCEPTION_REGION1; i++){
-      if (W->StartsWith(ExceptionsRegion1[i]))
-        return U32(strlen(ExceptionsRegion1[i]));
-    }
-    return GetRegion(W, 0);
-  }
-   
-  bool EndsInShortSyllable(const Word *W){
-    if (W->End==W->Start)
-      return false;
-    else if (W->End==W->Start+1)
-      return IsVowel((*W)(1)) && IsConsonant((*W)(0));
-    else
-      return (IsConsonant((*W)(2)) && IsVowel((*W)(1)) && IsConsonant((*W)(0)) && IsShortConsonant((*W)(0)));
-  }
-  bool IsShortWord(const Word *W){
-    return (EndsInShortSyllable(W) && GetRegion1(W)==W->Length());
-  }
-  inline bool HasVowels(const Word *W){
-    for (int i=W->Start; i<=W->End; i++){
-      if (IsVowel(W->Letters[i]))
-        return true;
-    }
-    return false;
-  }
-  bool TrimStartingApostrophe(Word *W){
-    bool result=false;
-    //trim all apostrophes from the beginning
-    int cnt=0;
-    while(W->Start!=W->End && (*W)[0]=='\'') {
-      result=true;
-      W->Start++;
-      cnt++;
-    }
-    //trim the same number of apostrophes from the end (if there are)
-    while(W->Start!=W->End && (*W)(0)=='\'') {
-      if(cnt==0)break;
-      W->End--;
-      cnt--;
-    }
-    return result;
-  }
-
-  void MarkYsAsConsonants(Word *W){
-    if ((*W)[0]=='y')
-      W->Letters[W->Start]='Y';
-    for (int i=W->Start+1; i<=W->End; i++){
-      if (IsVowel(W->Letters[i-1]) && W->Letters[i]=='y')
-        W->Letters[i]='Y';
-    }
-  }
-  bool ProcessPrefixes(Word *W){
-    if (W->StartsWith("irr") && W->Length()>5 && ((*W)[3]=='a' || (*W)[3]=='e'))
-      W->Start+=2, W->Type|=English::Negation;
-    else if (W->StartsWith("over") && W->Length()>5)
-      W->Start+=4, W->Type|=English::PrefixOver;
-    else if (W->StartsWith("under") && W->Length()>6)
-      W->Start+=5, W->Type|=English::PrefixUnder;
-    else if (W->StartsWith("unn") && W->Length()>5)
-      W->Start+=2, W->Type|=English::Negation;
-    else if (W->StartsWith("non") && W->Length()>(U32)(5+((*W)[3]=='-')))
-      W->Start+=2+((*W)[3]=='-'), W->Type|=English::Negation;
-    else
-      return false;
-    return true;
-  }
-  bool ProcessSuperlatives(Word *W){
-    if (W->EndsWith("est") && W->Length()>4){
-      U8 i=W->End;
-      W->End-=3;
-      W->Type|=English::AdjectiveSuperlative;
-
-      if ((*W)(0)==(*W)(1) && (*W)(0)!='r' && !(W->Length()>=4 && memcmp("sugg",&W->Letters[W->End-3],4)==0)){
-        W->End-= ( ((*W)(0)!='f' && (*W)(0)!='l' && (*W)(0)!='s') ||
-                   (W->Length()>4 && (*W)(1)=='l' && ((*W)(2)=='u' || (*W)(3)=='u' || (*W)(3)=='v'))) &&
-                   (!(W->Length()==3 && (*W)(1)=='d' && (*W)(2)=='o'));
-        if (W->Length()==2 && ((*W)[0]!='i' || (*W)[1]!='n'))
-          W->End = i, W->Type&=~English::AdjectiveSuperlative;
-      }
-      else{
-        switch((*W)(0)){
-          case 'd': case 'k': case 'm': case 'y': break;
-          case 'g': {
-            if (!( W->Length()>3 && ((*W)(1)=='n' || (*W)(1)=='r') && memcmp("cong",&W->Letters[W->End-3],4)!=0 ))
-              W->End = i, W->Type&=~English::AdjectiveSuperlative;
-            else
-              W->End+=((*W)(2)=='a');
-            break;
-          }
-          case 'i': {W->Letters[W->End]='y'; break;}
-          case 'l': {
-            if (W->End==W->Start+1 || memcmp("mo",&W->Letters[W->End-2],2)==0)
-              W->End = i, W->Type&=~English::AdjectiveSuperlative;
-            else
-              W->End+=IsConsonant((*W)(1));
-            break;
-          }
-          case 'n': {
-            if (W->Length()<3 || IsConsonant((*W)(1)) || IsConsonant((*W)(2)))
-              W->End = i, W->Type&=~English::AdjectiveSuperlative;
-            break;
-          }
-          case 'r': {
-            if (W->Length()>3 && IsVowel((*W)(1)) && IsVowel((*W)(2)))
-              W->End+=((*W)(2)=='u') && ((*W)(1)=='a' || (*W)(1)=='i');
-            else
-              W->End = i, W->Type&=~English::AdjectiveSuperlative;
-            break;
-          }
-          case 's': {W->End++; break;}
-          case 'w': {
-            if (!(W->Length()>2 && IsVowel((*W)(1))))
-              W->End = i, W->Type&=~English::AdjectiveSuperlative;
-            break;
-          }
-          case 'h': {
-            if (!(W->Length()>2 && IsConsonant((*W)(1))))
-              W->End = i, W->Type&=~English::AdjectiveSuperlative;
-            break;
-          }
-          default: {
-            W->End+=3;
-            W->Type&=~English::AdjectiveSuperlative;
-          }
-        }
-      }
-    }
-    return (W->Type&English::AdjectiveSuperlative)>0;
-  }
-  bool Step0(Word *W){
-    for (int i=0; i<NUM_SUFFIXES_STEP0; i++){
-      if (W->EndsWith(SuffixesStep0[i])){
-        W->End-=U8(strlen(SuffixesStep0[i]));
-        W->Type|=English::Plural;
-        return true;
-      }
-    }
-    return false;
-  }
-  bool Step1a(Word *W){
-    if (W->EndsWith("sses")){
-      W->End-=2;
-      W->Type|=English::Plural;
-      return true;
-    }
-    if (W->EndsWith("ied") || W->EndsWith("ies")){
-      W->Type|=((*W)(0)=='d')?English::PastTense:English::Plural;
-      W->End-= W->Length()>4 ? 2 : 1;
-      return true;
-    }
-    if (W->EndsWith("us") || W->EndsWith("ss"))
-      return false;
-    if ((*W)(0)=='s' && W->Length()>2){
-      for (int i=W->Start;i<=W->End-2;i++){
-        if (IsVowel(W->Letters[i])){
-          W->End--;
-          W->Type|=English::Plural;
-          return true;
-        }
-      }
-    }
-    if (W->EndsWith("n't") && W->Length()>4){
-      switch ((*W)(3)){
-        case 'a': {
-          if ((*W)(4)=='c')
-            W->End-=2;
-          else
-            W->ChangeSuffix("n't","ll");
-          break;
-        }
-        case 'i': {W->ChangeSuffix("in't","m"); break;}
-        case 'o': {
-          if ((*W)(4)=='w')
-            W->ChangeSuffix("on't","ill");
-          else
-            W->End-=3;
-          break;
-        }
-        default: W->End-=3;
-      }
-      W->Type|=English::Negation;
-      return true;
-    }
-    if (W->EndsWith("hood") && W->Length()>7){
-      W->End-=4;
-      return true;
-    }
-    return false;
-  }
-  bool Step1b(Word *W, const U32 R1){
-    for (int i=0; i<NUM_SUFFIXES_STEP1b; i++){
-      if (W->EndsWith(SuffixesStep1b[i])){
-        switch(i){
-          case 0: case 1: {
-            if (SuffixInRn(W, R1, SuffixesStep1b[i]))
-              W->End-=1+i*2;
-            break;
-          }
-          default: {
-            U8 j=W->End;
-            W->End-=U8(strlen(SuffixesStep1b[i]));
-            if (HasVowels(W)){
-              if (W->EndsWith("at") || W->EndsWith("bl") || W->EndsWith("iz") || IsShortWord(W))
-                (*W)+='e';
-              else if (W->Length()>2){
-                if ((*W)(0)==(*W)(1) && IsDouble((*W)(0)))
-                  W->End--;
-                else if (i==2 || i==3){
-                  switch((*W)(0)){
-                    case 'c': case 's': case 'v': {W->End+=!(W->EndsWith("ss") || W->EndsWith("ias")); break;}
-                    case 'd': {
-                      static const char nAllowed[4] = {'a','e','i','o'};
-                      W->End+=IsVowel((*W)(1)) && (!CharInArray((*W)(2), nAllowed, 4)); break;
-                    }
-                    case 'k': {W->End+=W->EndsWith("uak"); break;}
-                    case 'l': {
-                      static const char Allowed1[10] = {'b','c','d','f','g','k','p','t','y','z'};
-                      static const char Allowed2[4] = {'a','i','o','u'};
-                      W->End+= CharInArray((*W)(1), Allowed1, 10) ||
-                                (CharInArray((*W)(1), Allowed2, 4) && IsConsonant((*W)(2)));
-                      break;
-                    }
-                  }
-                }
-                else if (i>=4){
-                  switch((*W)(0)){
-                    case 'd': {
-                      if (IsVowel((*W)(1)) && (*W)(2)!='a' && (*W)(2)!='e' && (*W)(2)!='o')
-                        (*W)+='e';
-                      break;
-                    }
-                    case 'g': {
-                      static const char Allowed[7] = {'a','d','e','i','l','r','u'};
-                      if (
-                        CharInArray((*W)(1), Allowed, 7) || (
-                         (*W)(1)=='n' && (
-                          (*W)(2)=='e' ||
-                          ((*W)(2)=='u' && (*W)(3)!='b' && (*W)(3)!='d') ||
-                          ((*W)(2)=='a' && ((*W)(3)=='r' || ((*W)(3)=='h' && (*W)(4)=='c'))) ||
-                          (W->EndsWith("ring") && ((*W)(4)=='c' || (*W)(4)=='f'))
-                         )
-                        ) 
-                      )
-                        (*W)+='e';
-                      break;
-                    }
-                    case 'l': {
-                      if (!((*W)(1)=='l' || (*W)(1)=='r' || (*W)(1)=='w' || (IsVowel((*W)(1)) && IsVowel((*W)(2)))))
-                        (*W)+='e';
-                      if (W->EndsWith("uell") && W->Length()>4 && (*W)(4)!='q')
-                        W->End--;
-                      break;
-                    }
-                    case 'r': {
-                      if ((
-                        ((*W)(1)=='i' && (*W)(2)!='a' && (*W)(2)!='e' && (*W)(2)!='o') ||
-                        ((*W)(1)=='a' && (!((*W)(2)=='e' || (*W)(2)=='o' || ((*W)(2)=='l' && (*W)(3)=='l')))) ||
-                        ((*W)(1)=='o' && (!((*W)(2)=='o' || ((*W)(2)=='t' && (*W)(3)!='s')))) ||
-                        (*W)(1)=='c' || (*W)(1)=='t') && (!W->EndsWith("str"))
-                      )
-                        (*W)+='e';
-                      break;
-                    }
-                    case 't': {
-                      if ((*W)(1)=='o' && (*W)(2)!='g' && (*W)(2)!='l' && (*W)(2)!='i' && (*W)(2)!='o')
-                        (*W)+='e';
-                      break;
-                    }
-                    case 'u': {
-                      if (!(W->Length()>3 && IsVowel((*W)(1)) && IsVowel((*W)(2))))
-                        (*W)+='e';
-                      break;
-                    }
-                    case 'z': {
-                      if (W->EndsWith("izz") && W->Length()>3 && ((*W)(3)=='h' || (*W)(3)=='u'))
-                        W->End--;
-                      else if ((*W)(1)!='t' && (*W)(1)!='z')
-                        (*W)+='e';
-                      break;
-                    }
-                    case 'k': {
-                      if (W->EndsWith("uak"))
-                        (*W)+='e';
-                      break;
-                    }
-                    case 'b': case 'c': case 's': case 'v': {
-                      if (!(
-                        ((*W)(0)=='b' && ((*W)(1)=='m' || (*W)(1)=='r')) ||
-                        W->EndsWith("ss") || W->EndsWith("ias") || (*W)=="zinc"
-                      ))
-                        (*W)+='e';
-                      break;
-                    }
-                  }
-                }
-              }
-            }
-            else{
-              W->End=j;
-              return false;
-            }
-          }
-        }
-        W->Type|=TypesStep1b[i];
-        return true;
-      }
-    }
-    return false;
-  }
-  bool Step1c(Word *W){
-    if (W->Length()>2 && tolower((*W)(0))=='y' && IsConsonant((*W)(1))){
-      W->Letters[W->End]='i';
-      return true;
-    }
-    return false;
-  }
-  bool Step2(Word *W, const U32 R1){
-    for (int i=0; i<NUM_SUFFIXES_STEP2; i++){
-      if (W->EndsWith(SuffixesStep2[i][0]) && SuffixInRn(W, R1, SuffixesStep2[i][0])){
-        W->ChangeSuffix(SuffixesStep2[i][0], SuffixesStep2[i][1]);
-        W->Type|=TypesStep2[i];
-        return true;
-      }
-    }
-    if (W->EndsWith("logi") && SuffixInRn(W, R1, "ogi")){
-      W->End--;
-      return true;
-    }
-    else if (W->EndsWith("li")){
-      if (SuffixInRn(W, R1, "li") && IsLiEnding((*W)(2))){
-        W->End-=2;
-        W->Type|=English::AdverbOfManner;
-        return true;
-      }
-      else if (W->Length()>3){
-        switch((*W)(2)){
-            case 'b': {
-              W->Letters[W->End]='e';
-              W->Type|=English::AdverbOfManner;
-              return true;              
-            }
-            case 'i': {
-              if (W->Length()>4){
-                W->End-=2;
-                W->Type|=English::AdverbOfManner;
-                return true;
-              }
-              break;
-            }
-            case 'l': {
-              if (W->Length()>5 && ((*W)(3)=='a' || (*W)(3)=='u')){
-                W->End-=2;
-                W->Type|=English::AdverbOfManner;
-                return true;
-              }
-              break;
-            }
-            case 's': {
-              W->End-=2;
-              W->Type|=English::AdverbOfManner;
-              return true;
-            }
-            case 'e': case 'g': case 'm': case 'n': case 'r': case 'w': {
-              if (W->Length()>(U32)(4+((*W)(2)=='r'))){
-                W->End-=2;
-                W->Type|=English::AdverbOfManner;
-                return true;
-              }
-            }
-        }
-      }
-    }
-    return false;
-  }
-  bool Step3(Word *W, const U32 R1, const U32 R2){
-    bool res=false;
-    for (int i=0; i<NUM_SUFFIXES_STEP3; i++){
-      if (W->EndsWith(SuffixesStep3[i][0]) && SuffixInRn(W, R1, SuffixesStep3[i][0])){
-        W->ChangeSuffix(SuffixesStep3[i][0], SuffixesStep3[i][1]);
-        W->Type|=TypesStep3[i];
-        res=true;
-        break;
-      }
-    }
-    if (W->EndsWith("ative") && SuffixInRn(W, R2, "ative")){
-      W->End-=5;
-      W->Type|=English::SuffixIVE;
-      return true;
-    }
-    if (W->Length()>5 && W->EndsWith("less")){
-      W->End-=4;
-      W->Type|=English::AdjectiveWithout;
-      return true;
-    }
-    return res;
-  }
-  bool Step4(Word *W, const U32 R2){
-    bool res=false;
-    for (int i=0; i<NUM_SUFFIXES_STEP4; i++){
-      if (W->EndsWith(SuffixesStep4[i]) && SuffixInRn(W, R2, SuffixesStep4[i])){
-        W->End-=U8(strlen(SuffixesStep4[i])-(i>17));
-        if (!(i==10 /* ent */ && (*W)(0)=='m'))
-          W->Type|=TypesStep4[i];
-        if (i==0 && W->EndsWith("nti")){
-          W->End--;
-          res=true;
-          continue;
-        }
-        return true;
-      }
-    }
-    return res;
-  }
-  bool Step5(Word *W, const U32 R1, const U32 R2){
-    if ((*W)(0)=='e' && (*W)!="here"){
-      if (SuffixInRn(W, R2, "e"))
-        W->End--;
-      else if (SuffixInRn(W, R1, "e")){
-        W->End--;
-        W->End+=!EndsInShortSyllable(W);
-      }
-      else
-        return false;
-      return true;
-    }
-    else if (W->Length()>1 && (*W)(0)=='l' && SuffixInRn(W, R2, "l") && (*W)(1)=='l'){
-      W->End--;
-      return true;
-    }
-    return false;
-  }
-public:
-  inline bool IsVowel(const char c) final {
-    return CharInArray(c, Vowels, NUM_VOWELS);
-  }
-  inline void Hash(Word *W) final {
-    W->Hash[2] = W->Hash[3] = 0xb0a710ad;
-    for (int i=W->Start; i<=W->End; i++) {
-      U8 l = W->Letters[i];
-      W->Hash[2]=W->Hash[2]*263*32 + l;
-      if (IsVowel(l))
-        W->Hash[3]=W->Hash[3]*997*8 + (l/4-22);
-      else if (l>='b' && l<='z')
-        W->Hash[3]=W->Hash[3]*271*32 + (l-97);
-      else
-        W->Hash[3]=W->Hash[3]*11*32 + l;
-    }
-  }
-  bool Stem(Word *W){
-    if (W->Length()<2){
-      Hash(W);
-      return false;
-    }
-    bool res = TrimStartingApostrophe(W);
-    res|=ProcessPrefixes(W);
-    res|=ProcessSuperlatives(W);
-    for (int i=0; i<NUM_EXCEPTIONS1; i++){
-      if ((*W)==Exceptions1[i][0]){
-        if (i<11){
-          size_t len=strlen(Exceptions1[i][1]);
-          memcpy(&W->Letters[W->Start], Exceptions1[i][1], len);
-          W->End=W->Start+U8(len-1);
-        }
-        Hash(W);
-        W->Type|=TypesExceptions1[i];
-        W->Language = Language::English;
-        return (i<11);
-      }
-    }
-
-    // Start of modified Porter2 Stemmer
-    MarkYsAsConsonants(W);
-    U32 R1=GetRegion1(W), R2=GetRegion(W,R1);
-    res|=Step0(W);
-    res|=Step1a(W);
-    for (int i=0; i<NUM_EXCEPTIONS2; i++){
-      if ((*W)==Exceptions2[i]){
-        Hash(W);
-        W->Type|=TypesExceptions2[i];
-        W->Language = Language::English;
-        return res;
-      }
-    }
-    res|=Step1b(W, R1);
-    res|=Step1c(W);
-    res|=Step2(W, R1);
-    res|=Step3(W, R1, R2);
-    res|=Step4(W, R2);
-    res|=Step5(W, R1, R2);
-
-    for (U8 i=W->Start; i<=W->End; i++){
-      if (W->Letters[i]=='Y')
-        W->Letters[i]='y';
-    }
-    if (!W->Type || W->Type==English::Plural) {
-      if (W->MatchesAny(MaleWords, NUM_MALE_WORDS))
-        res = true, W->Type|=English::Male;
-      else if (W->MatchesAny(FemaleWords, NUM_FEMALE_WORDS))
-        res = true, W->Type|=English::Female;
-    }
-    if (!res)
-      res=W->MatchesAny(CommonWords, NUM_COMMON_WORDS);
-    Hash(W);
-    if (res)
-      W->Language = Language::English;
-    return res;
-  }
-};
-
-
-class FrenchStemmer: public Stemmer {
-private:
-  static const int NUM_VOWELS = 17;
-  const char Vowels[NUM_VOWELS]={'a','e','i','o','u','y','\xE2','\xE0','\xEB','\xE9','\xEA','\xE8','\xEF','\xEE','\xF4','\xFB','\xF9'};
-  static const int NUM_COMMON_WORDS = 10;
-  const char *CommonWords[NUM_COMMON_WORDS]={"de","la","le","et","en","un","une","du","que","pas"};
-  static const int NUM_EXCEPTIONS = 3;
-  const char *(Exceptions[NUM_EXCEPTIONS])[2]={
-    {"monument", "monument"},
-    {"yeux", "oeil"},
-    {"travaux", "travail"},
-  };
-  const U32 TypesExceptions[NUM_EXCEPTIONS]={
-    French::Noun,
-    French::Noun|French::Plural,
-    French::Noun|French::Plural
-  };
-  static const int NUM_SUFFIXES_STEP1 = 39;
-  const char *SuffixesStep1[NUM_SUFFIXES_STEP1]={
-    "ance","iqUe","isme","able","iste","eux","ances","iqUes","ismes","ables","istes", //11
-    "atrice","ateur","ation","atrices","ateurs","ations", //6
-    "logie","logies", //2
-    "usion","ution","usions","utions", //4
-    "ence","ences", //2
-    "issement","issements", //2
-    "ement","ements", //2
-    "it\xE9","it\xE9s", //2
-    "if","ive","ifs","ives", //4
-    "euse","euses", //2
-    "ment","ments" //2
-  };
-  static const int NUM_SUFFIXES_STEP2a = 35;
-  const char *SuffixesStep2a[NUM_SUFFIXES_STEP2a]={
-    "issaIent", "issantes", "iraIent", "issante",
-    "issants", "issions", "irions", "issais",
-    "issait", "issant", "issent", "issiez", "issons",
-    "irais", "irait", "irent", "iriez", "irons",
-    "iront", "isses", "issez", "\xEEmes",
-    "\xEEtes", "irai", "iras", "irez", "isse",
-    "ies", "ira", "\xEEt", "ie", "ir", "is",
-    "it", "i"
-  };
-  static const int NUM_SUFFIXES_STEP2b = 38;
-  const char *SuffixesStep2b[NUM_SUFFIXES_STEP2b]={
-    "eraIent", "assions", "erions", "assent",
-    "assiez", "\xE8rent", "erais", "erait",
-    "eriez", "erons", "eront", "aIent", "antes",
-    "asses", "ions", "erai", "eras", "erez",
-    "\xE2mes", "\xE2tes", "ante", "ants",
-    "asse", "\xE9""es", "era", "iez", "ais",
-    "ait", "ant", "\xE9""e", "\xE9s", "er",
-    "ez", "\xE2t", "ai", "as", "\xE9", "a"
-  };
-  static const int NUM_SET_STEP4 = 6;
-  const char SetStep4[NUM_SET_STEP4]={'a','i','o','u','\xE8','s'};
-  static const int NUM_SUFFIXES_STEP4 = 7;
-  const char *SuffixesStep4[NUM_SUFFIXES_STEP4]={"i\xE8re","I\xE8re","ion","ier","Ier","e","\xEB"};
-  static const int NUM_SUFFIXES_STEP5 = 5;
-  const char *SuffixesStep5[NUM_SUFFIXES_STEP5]={"enn","onn","ett","ell","eill"};
-  inline bool IsConsonant(const char c){
-    return !IsVowel(c);
-  }
-    void ConvertUTF8(Word *W) {
-    for (int i=W->Start; i<W->End; i++) {
-      U8 c = W->Letters[i+1]+((W->Letters[i+1]<0xA0)?0x60:0x40);
-      if (W->Letters[i]==0xC3 && (IsVowel(c) || (W->Letters[i+1]&0xDF)==0x87)) {
-        W->Letters[i] = c;
-        if (i+1<W->End)
-          memmove(&W->Letters[i+1], &W->Letters[i+2], W->End-i-1);
-        W->End--;
-      }
-    }
-  }
-  void MarkVowelsAsConsonants(Word *W){
-    for (int i=W->Start; i<=W->End; i++){
-      switch (W->Letters[i]) {
-        case 'i': case 'u': {
-          if (i>W->Start && i<W->End && (IsVowel(W->Letters[i-1]) || (W->Letters[i-1]=='q' && W->Letters[i]=='u')) && IsVowel(W->Letters[i+1]))
-            W->Letters[i] = toupper(W->Letters[i]);
-          break;
-        }
-        case 'y': {
-          if ((i>W->Start && IsVowel(W->Letters[i-1])) || (i<W->End && IsVowel(W->Letters[i+1])))
-            W->Letters[i] = toupper(W->Letters[i]);
-        }
-      }
-    }
-  }
-  U32 GetRV(Word *W) {
-    U32 len = W->Length(), res = W->Start+len;
-    if (len>=3 && ((IsVowel(W->Letters[W->Start]) && IsVowel(W->Letters[W->Start+1])) || W->StartsWith("par") || W->StartsWith("col") || W->StartsWith("tap") ))
-      return W->Start+3;
-    else {
-      for (int i=W->Start+1;i<=W->End;i++) {
-        if (IsVowel(W->Letters[i]))
-          return i+1;
-      }
-    }
-    return res;
-  }
-     bool Step1(Word *W, const U32 RV, const U32 R1, const U32 R2, bool *ForceStep2a) {
-    int i = 0;
-    for (; i<11; i++) {
-      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R2, SuffixesStep1[i])) {
-        W->End-=U8(strlen(SuffixesStep1[i]));
-        if (i==3 /*able*/)
-          W->Type|=French::Adjective;
-        return true;
-      }
-    }
-    for (; i<17; i++) {
-      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R2, SuffixesStep1[i])) {
-        W->End-=U8(strlen(SuffixesStep1[i]));
-        if (W->EndsWith("ic"))
-          W->ChangeSuffix("c", "qU");
-        return true;
-      }
-    }
-    for (; i<25;i++) {
-      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R2, SuffixesStep1[i])) {
-        W->End-=U8(strlen(SuffixesStep1[i]))-1-(i<19)*2;
-        if (i>22) {
-          W->End+=2;
-          W->Letters[W->End]='t';
-        }
-        return true;
-      }
-    }
-    for (; i<27; i++) {
-      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R1, SuffixesStep1[i]) && IsConsonant((*W)((U8)strlen(SuffixesStep1[i])))) {
-        W->End-=U8(strlen(SuffixesStep1[i]));
-        return true;
-      }
-    }
-    for (; i<29; i++) {
-      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, RV, SuffixesStep1[i])) {
-        W->End-=U8(strlen(SuffixesStep1[i]));
-        if (W->EndsWith("iv") && SuffixInRn(W, R2, "iv")) {
-          W->End-=2;
-          if (W->EndsWith("at") && SuffixInRn(W, R2, "at"))
-            W->End-=2;
-        }
-        else if (W->EndsWith("eus")) {
-          if (SuffixInRn(W, R2, "eus"))
-            W->End-=3;
-          else if (SuffixInRn(W, R1, "eus"))
-            W->Letters[W->End]='x';
-        }
-        else if ((W->EndsWith("abl") && SuffixInRn(W, R2, "abl")) || (W->EndsWith("iqU") && SuffixInRn(W, R2, "iqU")))
-          W->End-=3;
-        else if ((W->EndsWith("i\xE8r") && SuffixInRn(W, RV, "i\xE8r")) || (W->EndsWith("I\xE8r") && SuffixInRn(W, RV, "I\xE8r"))) {
-          W->End-=2;
-          W->Letters[W->End]='i';
-        }
-        return true;
-      }
-    }
-    for (; i<31; i++) {
-      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R2, SuffixesStep1[i])) {
-        W->End-=U8(strlen(SuffixesStep1[i]));
-        if (W->EndsWith("abil")) {
-          if (SuffixInRn(W, R2, "abil"))
-            W->End-=4;
-          else
-            W->End--, W->Letters[W->End]='l';
-        }
-        else if (W->EndsWith("ic")) {
-          if (SuffixInRn(W, R2, "ic"))
-            W->End-=2;
-          else
-            W->ChangeSuffix("c", "qU");
-        }
-        else if (W->EndsWith("iv") && SuffixInRn(W, R2, "iv"))
-          W->End-=2;
-        return true;
-      }
-    }
-    for (; i<35; i++) {
-      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R2, SuffixesStep1[i])) {
-        W->End-=U8(strlen(SuffixesStep1[i]));
-        if (W->EndsWith("at") && SuffixInRn(W, R2, "at")) {
-          W->End-=2;
-          if (W->EndsWith("ic")) {
-            if (SuffixInRn(W, R2, "ic"))
-              W->End-=2;
-            else
-              W->ChangeSuffix("c", "qU");
-          }
-        }
-        return true;
-      }
-    }
-    for (; i<37; i++) {
-      if (W->EndsWith(SuffixesStep1[i])) {
-        if (SuffixInRn(W, R2, SuffixesStep1[i])) {
-          W->End-=U8(strlen(SuffixesStep1[i]));
-          return true;
-        }
-        else if (SuffixInRn(W, R1, SuffixesStep1[i])) {
-          W->ChangeSuffix(SuffixesStep1[i], "eux");
-          return true;
-        }
-      }
-    }
-    for (; i<NUM_SUFFIXES_STEP1; i++) {
-      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, RV+1, SuffixesStep1[i]) && IsVowel((*W)((U8)strlen(SuffixesStep1[i])))) {
-        W->End-=U8(strlen(SuffixesStep1[i]));
-        (*ForceStep2a) = true;
-        return true;
-      }
-    }
-    if (W->EndsWith("eaux") || (*W)=="eaux") {
-      W->End--;
-      W->Type|=French::Plural;
-      return true;
-    }
-    else if (W->EndsWith("aux") && SuffixInRn(W, R1, "aux")) {
-      W->End--, W->Letters[W->End] = 'l';
-      W->Type|=French::Plural;
-      return true;
-    }
-    else if (W->EndsWith("amment") && SuffixInRn(W, RV, "amment")) {
-      W->ChangeSuffix("amment", "ant");
-      (*ForceStep2a) = true;
-      return true;
-    }
-    else if (W->EndsWith("emment") && SuffixInRn(W, RV, "emment")) {
-      W->ChangeSuffix("emment", "ent");
-      (*ForceStep2a) = true;
-      return true;
-    }
-    return false;
-  }
-  bool Step2a(Word *W, const U32 RV) {
-    for (int i=0; i<NUM_SUFFIXES_STEP2a; i++) {
-      if (W->EndsWith(SuffixesStep2a[i]) && SuffixInRn(W, RV+1, SuffixesStep2a[i]) && IsConsonant((*W)((U8)strlen(SuffixesStep2a[i])))) {
-        W->End-=U8(strlen(SuffixesStep2a[i]));
-        if (i==31 /*ir*/)
-          W->Type|=French::Verb;
-        return true;
-      }
-    }
-    return false;
-  }
-  bool Step2b(Word *W, const U32 RV, const U32 R2) {
-    for (int i=0; i<NUM_SUFFIXES_STEP2b; i++) {
-      if (W->EndsWith(SuffixesStep2b[i]) && SuffixInRn(W, RV, SuffixesStep2b[i])) {
-        switch (SuffixesStep2b[i][0]) {
-          case 'a': case '\xE2': {
-            W->End-=U8(strlen(SuffixesStep2b[i]));
-            if (W->EndsWith("e") && SuffixInRn(W, RV, "e"))
-              W->End--;
-            return true;
-          }
-          default: {
-            if (i!=14 || SuffixInRn(W, R2, SuffixesStep2b[i])) {
-              W->End-=U8(strlen(SuffixesStep2b[i]));
-              return true;
-            }
-          }
-        }        
-      }
-    }
-    return false;
-  }
-  void Step3(Word *W) {
-    char *final = (char *)&W->Letters[W->End];
-    if ((*final)=='Y')
-      (*final) = 'i';
-    else if ((*final)=='\xE7')
-      (*final) = 'c';
-  }
-  bool Step4(Word *W, const U32 RV, const U32 R2) {
-    bool res = false;
-    if (W->Length()>=2 && W->Letters[W->End]=='s' && !CharInArray((*W)(1), SetStep4, NUM_SET_STEP4)) {
-      W->End--;
-      res = true;
-    }
-    for (int i=0; i<NUM_SUFFIXES_STEP4; i++) {
-      if (W->EndsWith(SuffixesStep4[i]) && SuffixInRn(W, RV, SuffixesStep4[i])) {
-        switch (i) {
-          case 2: { //ion
-            char prec = (*W)(3);
-            if (SuffixInRn(W, R2, SuffixesStep4[i]) && SuffixInRn(W, RV+1, SuffixesStep4[i]) && (prec=='s' || prec=='t')) {
-              W->End-=3;
-              return true;
-            }
-            break;
-          }
-          case 5: { //e
-            W->End--;
-            return true;
-          }
-          case 6: { //\xEB
-            if (W->EndsWith("gu\xEB")) {
-              W->End--;
-              return true;
-            }
-            break;
-          }
-          default: {
-            W->ChangeSuffix(SuffixesStep4[i], "i");
-            return true;
-          }
-        }
-      }
-    }
-    return res;
-  }
-  bool Step5(Word *W) {
-    for (int i=0; i<NUM_SUFFIXES_STEP5; i++) {
-      if (W->EndsWith(SuffixesStep5[i])) {
-        W->End--;
-        return true;
-      }
-    }
-    return false;
-  }
-  bool Step6(Word *W) {
-    for (int i=W->End; i>=W->Start; i--) {
-      if (IsVowel(W->Letters[i])) {
-        if (i<W->End && (W->Letters[i]&0xFE)==0xE8) {
-          W->Letters[i] = 'e';
-          return true;
-        }
-        return false;
-      }
-    }
-    return false;
-  }
-public:
-  inline bool IsVowel(const char c) final {
-    return CharInArray(c, Vowels, NUM_VOWELS);
-  }
-  inline void Hash(Word *W) final {
-    W->Hash[2] = W->Hash[3] = ~0xeff1cace;
-    for (int i=W->Start; i<=W->End; i++) {
-      U8 l = W->Letters[i];
-      W->Hash[2]=W->Hash[2]*251*32 + l;
-      if (IsVowel(l))
-        W->Hash[3]=W->Hash[3]*997*16 + l;
-      else if (l>='b' && l<='z')
-        W->Hash[3]=W->Hash[3]*271*32 + (l-97);
-      else
-        W->Hash[3]=W->Hash[3]*11*32 + l;
-    }
-  }
-  bool Stem(Word *W) {
-    ConvertUTF8(W);
-    if (W->Length()<2) {
-      Hash(W);
-      return false;
-    }
-    for (int i=0; i<NUM_EXCEPTIONS; i++) {
-      if ((*W)==Exceptions[i][0]) {
-        size_t len=strlen(Exceptions[i][1]);
-        memcpy(&W->Letters[W->Start], Exceptions[i][1], len);
-        W->End=W->Start+U8(len-1);
-        Hash(W);
-        W->Type|=TypesExceptions[i];
-        W->Language = Language::French;
-        return true;
-      }
-    }
-    MarkVowelsAsConsonants(W);
-    U32 RV=GetRV(W), R1=GetRegion(W, 0), R2=GetRegion(W, R1);
-    bool DoNextStep=false, res=Step1(W, RV, R1, R2, &DoNextStep);
-    DoNextStep|=!res;
-    if (DoNextStep) {
-      DoNextStep = !Step2a(W, RV);
-      res|=!DoNextStep;
-      if (DoNextStep)
-        res|=Step2b(W, RV, R2);
-    }
-    if (res)
-      Step3(W);
-    else
-      res|=Step4(W, RV, R2);
-    res|=Step5(W);
-    res|=Step6(W);
-    for (int i=W->Start; i<=W->End; i++)
-      W->Letters[i] = tolower(W->Letters[i]);
-    if (!res)
-      res=W->MatchesAny(CommonWords, NUM_COMMON_WORDS);
-    Hash(W);
-    if (res)
-      W->Language = Language::French;
-    return res;
-  }
-};
-
-class GermanStemmer : public Stemmer {
-private:
-  static const int NUM_VOWELS = 9;
-  const char Vowels[NUM_VOWELS]={'a','e','i','o','u','y','\xE4','\xF6','\xFC'};
-  static const int NUM_COMMON_WORDS = 10;
-  const char *CommonWords[NUM_COMMON_WORDS]={"der","die","das","und","sie","ich","mit","sich","auf","nicht"};
-  static const int NUM_ENDINGS = 10;
-  const char Endings[NUM_ENDINGS]={'b','d','f','g','h','k','l','m','n','t'}; //plus 'r' for words ending in 's'
-  static const int NUM_SUFFIXES_STEP1 = 6;
-  const char *SuffixesStep1[NUM_SUFFIXES_STEP1]={"em","ern","er","e","en","es"};
-  static const int NUM_SUFFIXES_STEP2 = 3;
-  const char *SuffixesStep2[NUM_SUFFIXES_STEP2]={"en","er","est"};
-  static const int NUM_SUFFIXES_STEP3 = 7;
-  const char *SuffixesStep3[NUM_SUFFIXES_STEP3]={"end","ung","ik","ig","isch","lich","heit"};
-  void ConvertUTF8(Word *W) {
-    for (int i=W->Start; i<W->End; i++) {
-      U8 c = W->Letters[i+1]+((W->Letters[i+1]<0x9F)?0x60:0x40);
-      if (W->Letters[i]==0xC3 && (IsVowel(c) || c==0xDF)) {
-        W->Letters[i] = c;
-        if (i+1<W->End)
-          memmove(&W->Letters[i+1], &W->Letters[i+2], W->End-i-1);
-        W->End--;
-      }
-    }
-  }
-  void ReplaceSharpS(Word *W) {
-    for (int i=W->Start; i<=W->End; i++) {
-      if (W->Letters[i]==0xDF) {
-        W->Letters[i]='s';
-        if (i+1<MAX_WORD_SIZE) {
-          memmove(&W->Letters[i+2], &W->Letters[i+1], MAX_WORD_SIZE-i-2);
-          W->Letters[i+1]='s';
-          W->End+=(W->End<MAX_WORD_SIZE-1);
-        }
-      }
-    }
-  }    
-  void MarkVowelsAsConsonants(Word *W) {
-    for (int i=W->Start+1; i<W->End; i++) {
-      U8 c = W->Letters[i];
-      if ((c=='u' || c=='y') && IsVowel(W->Letters[i-1]) && IsVowel(W->Letters[i+1]))
-        W->Letters[i] = toupper(c);
-    }
-  }
-  inline bool IsValidEnding(const char c, const bool IncludeR = false) {
-    return CharInArray(c, Endings, NUM_ENDINGS) || (IncludeR && c=='r');
-  }
-  bool Step1(Word *W, const U32 R1) {
-    int i = 0;
-    for (; i<3; i++) {
-      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R1, SuffixesStep1[i])) {
-        W->End-=U8(strlen(SuffixesStep1[i]));
-        return true;
-      }
-    }
-    for (; i<NUM_SUFFIXES_STEP1; i++) {
-      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R1, SuffixesStep1[i])) {
-        W->End-=U8(strlen(SuffixesStep1[i]));
-        W->End-=U8(W->EndsWith("niss"));
-        return true;
-      }
-    }
-    if (W->EndsWith("s") && SuffixInRn(W, R1, "s") && IsValidEnding((*W)(1), true)) {
-      W->End--;
-      return true;
-    }
-    return false;
-  }
-  bool Step2(Word *W, const U32 R1) {
-    for (int i=0; i<NUM_SUFFIXES_STEP2; i++) {
-      if (W->EndsWith(SuffixesStep2[i]) && SuffixInRn(W, R1, SuffixesStep2[i])) {
-        W->End-=U8(strlen(SuffixesStep2[i]));
-        return true;
-      }
-    }
-    if (W->EndsWith("st") && SuffixInRn(W, R1, "st") && W->Length()>5 && IsValidEnding((*W)(2))) {
-      W->End-=2;
-      return true;
-    }
-    return false;
-  }
-  bool Step3(Word *W, const U32 R1, const U32 R2) {
-    int i = 0;
-    for (; i<2; i++) {
-      if (W->EndsWith(SuffixesStep3[i]) && SuffixInRn(W, R2, SuffixesStep3[i])) {
-        W->End-=U8(strlen(SuffixesStep3[i]));
-        if (W->EndsWith("ig") && (*W)(2)!='e' && SuffixInRn(W, R2, "ig"))
-          W->End-=2;
-        if (i)
-          W->Type|=German::Noun;
-        return true;
-      }
-    }
-    for (; i<5; i++) {
-      if (W->EndsWith(SuffixesStep3[i]) && SuffixInRn(W, R2, SuffixesStep3[i]) && (*W)((U8)strlen(SuffixesStep3[i]))!='e') {
-        W->End-=U8(strlen(SuffixesStep3[i]));
-        if (i>2)
-          W->Type|=German::Adjective;
-        return true;
-      }
-    }
-    for (; i<NUM_SUFFIXES_STEP3; i++) {
-      if (W->EndsWith(SuffixesStep3[i]) && SuffixInRn(W, R2, SuffixesStep3[i])) {
-        W->End-=U8(strlen(SuffixesStep3[i]));
-        if ((W->EndsWith("er") || W->EndsWith("en")) && SuffixInRn(W, R1, "e?"))
-          W->End-=2;
-        if (i>5)
-          W->Type|=German::Noun|German::Female;
-        return true;
-      }
-    }
-    if (W->EndsWith("keit") && SuffixInRn(W, R2, "keit")) {
-      W->End-=4;
-      if (W->EndsWith("lich") && SuffixInRn(W, R2, "lich"))
-        W->End-=4;
-      else if (W->EndsWith("ig") && SuffixInRn(W, R2, "ig"))
-        W->End-=2;
-      W->Type|=German::Noun|German::Female;
-      return true;
-    }
-    return false;
-  }
-public:
-  inline bool IsVowel(const char c) final {
-    return CharInArray(c, Vowels, NUM_VOWELS);
-  }
-  inline void Hash(Word *W) final {
-    W->Hash[2] = W->Hash[3] = ~0xbea7ab1e;
-    for (int i=W->Start; i<=W->End; i++) {
-      U8 l = W->Letters[i];
-      W->Hash[2]=W->Hash[2]*263*32 + l;
-      if (IsVowel(l))
-        W->Hash[3]=W->Hash[3]*997*16 + l;
-      else if (l>='b' && l<='z')
-        W->Hash[3]=W->Hash[3]*251*32 + (l-97);
-      else
-        W->Hash[3]=W->Hash[3]*11*32 + l;
-    }
-  }
-  bool Stem(Word *W) {
-    ConvertUTF8(W);
-    if (W->Length()<2) {
-      Hash(W);
-      return false;
-    }
-    ReplaceSharpS(W);
-    MarkVowelsAsConsonants(W);
-    U32 R1=GetRegion(W, 0), R2=GetRegion(W, R1);
-    R1 = min(3, R1);
-    bool res = Step1(W, R1);
-    res|=Step2(W, R1);
-    res|=Step3(W, R1, R2);
-    for (int i=W->Start; i<=W->End; i++) {
-      switch (W->Letters[i]) {
-        case 0xE4: { W->Letters[i] = 'a'; break; }
-        case 0xF6: case 0xFC: { W->Letters[i]-=0x87; break; }
-        default: W->Letters[i] = tolower(W->Letters[i]);
-      }
-    }
-    if (!res)
-      res=W->MatchesAny(CommonWords, NUM_COMMON_WORDS);
-    Hash(W);
-    if (res)
-      W->Language = Language::German;
-    return res;
-  }
-};
 
 //////////////////////////// Models //////////////////////////////
 
@@ -2372,697 +898,6 @@ void TextModel::SetContexts(Buf& buffer,Mixer& mixer) {
 
 
 
-    
-
-//Based on XWRT 3.2 (29.10.2007) - XML compressor by P.Skibinski, inikep@gmail.com
-#include "wrton.cpp"
-//////////////////////////// wordModel /////////////////////////
-
-#define TAB 0x09
-#define NEW_LINE 0x0A
-#define CARRIAGE_RETURN 0x0D
-#define SPACE 0x20
-#define QUOTE 0x22
-#define APOSTROPHE 0x27
-
-// Model English text (words and columns/end of line)
-class wordModel1: public Model {
-   public:
-   BlockData& x;
-   Buf& buf;  
-private:
-   int N;
-   ContextMap cm;
-   ContextMap cm1;
-  class Info {
-   BlockData& x;
-   Buf& buf;
-    ContextMap  &cm;
-    U32 word0, word1, word2, word3, word4, word5,word6,word7;  // hashes
-    U32 wrdhsh;
-    U32 xword0,xword1,xword2,xword3,cword0,ccword,fword;
-    U32 number0, number1,hq;  // hashes
-    U32 text0,data0,type0;  // hash stream of letters
-    int nl3,nl2,nl1, nl;  // previous, current newline position
-    U32 mask,mask2;
-    Array<int> wpos;  // last position of word
-    Array<U16> wchk ;
-    Array<U32> inkeyw;
-    Array<U32> inkeyw1;
-    int w;
-    bool doXML;
-    U32 lastLetter,firstLetter, lastUpper, lastDigit,wordGap;
-    Array<Word> StemWords;
-    Word *cWord, *pWord;
-    EnglishStemmer StemmerEN;
-    int StemIndex,same,linespace;
-    int islink,istemplate;
-    bool is_letter, is_letter_pC, is_letter_ppC;
-    U32 numbers,numbercount;
-    U64 g_ascii_lo=0, g_ascii_hi=0; // group identifiers of 8+8 last characters
-    U8 c, pC, ppC;           // last char, previous char, char before the previous char (converted to lowearcase)
-    U32 c4;
-    U32 ccount,col1;
-    bool firstwasspace;
-    U8 opened;
-    U32 line0;
-    U32 iword0;
-    U32 lastSpace;
-    bool pdft;
-    U16 chk;
-    int above, above1, above2,fl,f;
-    U32 expr0, expr1, expr2, expr3, expr4,exprlen0,expr0chars,wchar1,wchar2,wrth;  // wordtoken hashes for expressions
-    int wcol,wnl2,wnl1,wnl,wabove,scount;
-    bool scountset;
-    U32 wpword1;
-  public:
-   
-    Info(BlockData& bd,U32 val, ContextMap  &contextmap):x(bd),buf(bd.buf), cm(contextmap),
-    word0(0),word1(0),word2(0), word3(0),word4(0),word5(0),word6(0),word7(0),
-    wrdhsh(0),
-    xword0(0),xword1(0),xword2(0),xword3(0),cword0(0),ccword(0),fword(0),
-    number0(0),number1(0),hq(0),
-    text0(0),data0(0),type0(0),nl3(-5),nl2(-4),nl1(-3), nl(-2),mask(0),mask2(0),wpos(0x10000),wchk(0x10000),inkeyw(0x10000),inkeyw1(0x10000),w(0),doXML(false),
-  lastLetter(0),firstLetter(0), lastUpper(0),lastDigit(0), wordGap(0) ,StemWords(4),StemIndex(0),same(0),linespace(0),islink(0),istemplate(0),numbers(0),numbercount(0),
-  g_ascii_lo(0), g_ascii_hi(0), c(0), pC(0), ppC(0),c4(0),ccount(0),col1(0),firstwasspace(false),opened(0),line0(0),
-  iword0(0),lastSpace(0),pdft(false),chk(0),above(0), above1(0), above2(0),fl(0),f(0),scount(0),scountset(false),wpword1(0) {
-      reset();
-    }
-    void reset() { 
-       cWord=&StemWords[0], pWord=&StemWords[3];expr0=expr1=expr2=expr3=expr4=exprlen0=expr0chars=wchar2=wrth=0;wchar1=1;
-       is_letter=is_letter_pC=is_letter_ppC=false;
-       
-       wcol=wabove=0,wnl2,wnl1=wnl=-1;
-    }
-    // forget it
-    void shrwords() { 
-        word1=word2;
-        word2=word3;
-        word3=word4;
-        word4=word5;
-        word5=word6;
-        word6=word7;
-        word7=0;
-    }
-    void killwords() { 
-    word1=word2=word2=word3=word3=word4=0;
-        word4=word5=word5=word6=word6=word7=0;
-        word7=0;
-    }
-    void process_char(const int is_extended_char,int val1=0,int val2=0) {
-        ppC=pC; pC=c;
-        is_letter_ppC=is_letter_pC; is_letter_pC=is_letter;
-        c=x.c4&255;
-        c4=(c4<<8)|c;
-        int i=0,d=0;
-        if (f) {
-            word5=word4;
-            word4=word3;
-            word3=word2;
-            word2=word1;
-            word1='.';
-        }
-        f=0;
-        if (x.spaces&0x80000000) --x.spacecount;
-        if (x.words&0x80000000) --x.wordcount;
-        if (numbers&0x80000000) --numbercount;
-        numbers=numbers*2;
-        x.spaces=x.spaces*2;
-        x.words=x.words*2;
-        lastUpper=min(lastUpper+1,255);
-        lastLetter=min(lastLetter+1,255);
-        lastSpace=min(lastSpace+1,1024*4);
-        mask2<<=2;
-        if (is_extended_char>0 || is_extended_char==-1) wchar1=(wchar1<<8)|c;
-        if (is_extended_char==0 && wchar1!=1 ) wchar2=wchar1,wchar1=1;
-        if (!(x.filetype==DEFAULT || x.filetype==DECA || x.filetype==IMGUNK || x.filetype==EXE || x.filetype==ARM)){
-           d=WRT_wrd1[c]; 
-           x.pwords=x.pwords*2;   
-           x.pbc=x.pbc*2+d/64;
-           if ((d)) {x.pwords++, wpword1=(wpword1<<8)+c;}//hash(wpword1, c);}
-           else wpword1=d; 
-          // x.bc4=d<<2; 
-        }
-        if (pC>='A' && pC<='Z') pC+='a'-'A';
-        if (c>='A' && c<='Z') c+='a'-'A', lastUpper=0;
-        if (x.wrtLoaded==true && c==1 ) lastUpper=0;
-        if ((c>='a' && c<='z') || c=='\'' || c=='-')
-           (*cWord)+=c;
-        else if ((*cWord).Length()>0){
-           StemmerEN.Stem(cWord);
-           StemIndex=(StemIndex+1)&3;
-           pWord=cWord;
-           cWord=&StemWords[StemIndex];
-           memset(cWord, 0, sizeof(Word));
-        }
-        if ((c4&0xFFFF)==0x5b5b ) scountset=true;
-        if (islink && (c==SPACE || c==']' || c==10  )) islink=0; //disable if not in link
-        if (scountset &&  c=='|')  scountset=false; 
-        if ((c4&0xFFFF)==0x5d5d && cword0==-1)cword0=0;//']]'
-        if (scountset &&  (c==']' ||x.frstchar=='[') )  scountset=false,scount=cword0=0; 
-        if (istemplate && (c4&0xffff)==0x7d7d) istemplate=0; //'}}'
-        if ((val1==0 || val1==1)&& doXML==true) doXML=false; // None ReadTag
-        else if (val1==5) doXML=true;                        // ReadContent
-        //if (doXML==true) printf("%c",c);
-        is_letter=((c>='a' && c<='z') ||(c>='0' && c<='9' && witmode==true) || x.wstat==true|| (c>=128 &&(x.b3!=3)|| (c>0 && c<4 )));
-        if (is_letter) {// if ((c>='a' && c<='z')||/*(c>='0' && c<='9') ||*/ (c>=128 /*&&(x.b3!=3)*/|| (c>0 && c<4))) {
-            if (!x.wordlen){
-                // model syllabification with:
-                //       "+"  //book1 case +\n +\r\n
-                //       "="  quoted-printable 
-                if ((lastLetter=3 && (c4&0xFFFF00)==0x2B0A00 && buf(4)!=0x2B) || (lastLetter=4 && (c4&0xFFFFFF00)==0x2B0D0A00 && buf(5)!=0x2B) ||
-                    (lastLetter=3 && (c4&0xFFFF00)==0x2D0A00 && buf(4)!=0x2D) || (lastLetter=4 && (c4&0xFFFFFF00)==0x2D0D0A00 && buf(5)!=0x2D) ||
-                    (lastLetter=3 && (c4&0xFFFF00)==0x3D0A00 && buf(4)!=0x3D) || (lastLetter=4 && (c4&0xFFFFFF00)==0x3D0D0A00 && buf(5)!=0x3D)  ){
-                    word0=word1;
-                    shrwords();
-                    x.wordlen = x.wordlen1;
-                    if (c<128){
-                       StemIndex=(StemIndex-1)&3;
-                       cWord=pWord;
-                       pWord=&StemWords[(StemIndex-1)&3];
-                       memset(cWord, 0, sizeof(Word));
-                       for (U32 j=0;j<=x.wordlen;j++)
-                           (*cWord)+=tolower(buf(x.wordlen-j+1+2*(j!=x.wordlen)));
-                    }
-                }else{
-                      wordGap = lastLetter;
-                      firstLetter = c;
-                      wrdhsh = 0;
-                }
-                if (pC==QUOTE /*||   (x.wrtc4&0xffffffffffff)==0x2671756F743B*/) { opened=QUOTE;}
-            }
-            lastLetter=0;
-            if (c>4) word0^=hash(word0, c,0);
-            text0=text0*997*16+c;
-            ++x.words, ++x.wordcount;
-            x.wordlen++;
-            x.wordlen=min(x.wordlen,45);
-            if (x.wstat==true && x.wdecoded==true) wrth=wrth+x.utf8l>0;
-        
-        if (x.wstat==false ) wrth++;
-            if ((c=='a' || c=='e' || c=='i' || c=='o' || c=='u') || (c=='y' && (x.wordlen>0 && pC!='a' && pC!='e' && pC!='i' && pC!='o' && pC!='u'))){
-                mask2++;
-                wrdhsh=wrdhsh*997*8+(c/4-22);
-            }else if (c>='b' && c<='z'){
-                mask2+=2;
-                wrdhsh=wrdhsh*271*32+(c-97);
-            }else
-                wrdhsh=wrdhsh*11*32+c;
-            f=0;
-            w=word0&(wpos.size()-1);
-            chk=U16(word0>>16);
-        } else {
-            if (word0) { 
-                type0 = (type0<<2)|1;
-                if (scountset) scount=min(scount+1,5);
-                if (scountset==false && scount>0&& cword0!=-1) { // wiki: xxx xxx [[yyy yyy|xxx xxx]] -> remove yyy words
-                    x.f4=x.f4>>(4*scount);
-                    x.tt=x.tt>>(2*scount);
-                    while (scount-->0) shrwords();
-                    scount=0;
-                }
-                else if (cword0==-1) { // wiki: xxx xxx [[yyy: yyy|yyy yyy]] -> remove yyy words
-                    x.f4=x.f4>>(4*1);
-                    x.tt=x.tt>>(2*1);
-                    shrwords();
-                    scount=0;
-                }else{ 
-                    word7=word6;
-                    word6=word5;     
-                    word5=word4;
-                    word4=word3;
-                    word3=word2;
-                    word2=word1;
-                    word1=word0;
-                    x.wordlen1=x.wordlen;
-                }
-                x.wordlen=0;
-                inkeyw[w]=iword0;
-                if (fword==0) fword=word0;
-                
-                wpos[w]=x.blpos;
-                wrth=0;
-                wchk[w]=chk;
-                if (c==':'|| c=='=') cword0=word0;
-                if ((c==':' && scountset==true && scount==1)) cword0=-1,shrwords(); // kill  [[xxx: ...]]
-                if (c==']'&& (x.frstchar!=':') &&  doXML==true) xword0=word0; // wiki 
-                ccword=0;
-                word0= 0;                
-                if((c=='.'||c=='!'||c=='?' ||c=='}' ||c==')') && buf(2)!=10 && x.filetype!=EXE) f=1; 
-            }
-            if (c==SPACE &&  buf(2)!=SPACE) linespace++; // count spaces in one line, skip repeats
-            if (c==SPACE || c==10 || (c==5 && pC!='&')) { 
-              if (c==SPACE )lastSpace=0,++x.spaces, ++x.spacecount; //else  
-              if (pC=='.' || (ppC=='.' &&( pC==0x27 || pC==0x22))  ){
-              x.tt=0xfffffff8, x.f4=0;
-              } else   if  (pC=='?' || ((ppC=='?'|| ppC=='!')&&( pC==0x27 || pC==0x22))  ){
-              x.tt=(x.tt&0xfffffff8)+1, x.f4=0;
-              } 
-               if (/*c==10 &&*/ linespace==0 && x.frstchar==0x5B ){
-                   xword1=xword2=xword3=iword0=0;
-               }
-               if (c==10 || (c==5 && pC!='&')) fword=linespace=line0=0,nl3=nl2,nl2=nl1,nl1=nl, nl=buf.pos-1,wnl2=wnl1,wnl1=wnl,wnl=x.bufn.pos-1;
-            }
-            else if ((x.wrtc4&0xffffffff)==0x266C743B ||
-                            (x.wrtc4&0xffffffff)==0x2667743B ||
-                    (x.wrtc4&0xffffffffffff)==0x2671756F743B ||
-                    (x.wrtc4&    0xffffffffff)==0x26616D703B ||
-                                 x.wrtc4==0x26616D703B6C743B ||
-                                 x.wrtc4==0x26616D703B67743B ||
-                                 x.wrtc4==0x6D703B6E6273703B ||
-                                 x.wrtc4==0x6D703B71756F743B ||
-                                 x.wrtc4==0x703B6E646173683B ||
-                                 x.wrtc4==0x703B6D646173683B) { 
-                shrwords(); 
-       
-            }//'&lt;' '&gt;' " & 
-            else if (c=='.' || c=='!' || c=='?' || c==',' || c==';' || c==':' || c=='|'|| c=='=') x.spafdo=line0=0,ccword=c,mask2+=3;
-            else { ++x.spafdo; x.spafdo=min(63,x.spafdo); }
-            if (ccword=='=' || (ccword==':' && doXML==false)) iword0=word1;
-        }
-        if ((c4&0xffffff)==0x3A2F2F) { // ://
-            islink=1,word2=13;
-        }
-        if ((c4&0xffff)==0x7b7b) { //'{{'
-            istemplate=1;
-        }
-        if (doXML==true){
-            if ((c4&0xFFFF)==0x3D3D && x.frstchar==0x3d) xword1=word1; // ,xword2=word2; // == wiki
-            if ((c4&0xFFFF)==0x2727) xword2=word1;                     //,xword2=word2; // '' wiki
-            if ((c4&0xFFFF)==0x7D7D) xword3=word1;                     //}} wiki
-        }
-        lastDigit=min(0xFF,lastDigit+1);
-        if (c>='0' && c<='9') {
-            ++numbers, ++numbercount;
-            if (buf(3)>='0' && buf(3)<='9' && (buf(2)=='.'||buf(2)==',')&& number0==0) {number0=number1; number1=0;} // 0.4645
-            number0^=hash(number0, c,1);
-            lastDigit = 0;
-        }
-        else if (number0) {
-            type0 = (type0<<2)|2;
-            number1=number0;
-            number0=0,ccword=0;
-        }
-        line0^=hash(line0, buf(1));
-        if (!((c>='a' && c<='z') ||(c>='0' && c<='9') || (c>=128 ))){
-            data0^=hash(data0, c,1);
-        }else if (data0) {
-            type0 = (type0<<2)|3;
-            data0=0;
-            }
-        col1=x.col;
-        x.col=min(255, buf.pos-nl);
-        if (x.dictonline==true){
-        
-        x.wcol=wcol= x.bufn.pos-wnl; // (wrt mode)
-        if (wcol<0) x.wcol=wcol=2;
-        wabove=x.bufn[wnl1+wcol];  // text column context, first (wrt mode)
-        int lasllen=(wnl-wnl1);
-        if (wcol>lasllen)wabove=0;
-        }
-        else x.wcol=wabove=0;
-         above=buf[nl1+x.col];if (x.col>(nl-nl1)) above=0; // text column context, first
-         above1=buf[nl2+x.col];if (x.col>(nl1-nl2)) above1=0; // text column context, second
-         above2=buf[nl3+x.col];if (x.col>(nl2-nl3)) above2=0; // text column context, 3
-        if (val2) x.col=val2,above=buf[buf.pos-x.col],above1=buf[buf.pos-x.col*2];;
-        if (x.col<=2) {x.frstchar=(x.col==2?min(c,128):0); firstwasspace=x.frstchar==' '?true:false;        }
-        if (x.col>2 && firstwasspace && (x.frstchar==' ' || x.frstchar<4)  && buf(1)!=' ') x.frstchar=c;
-        if (x.col>2 && firstwasspace && !((x.frstchar>='a' && x.frstchar<='z') || (x.frstchar>='0' && x.frstchar<='9')|| (x.frstchar>=128 &&(x.b3!=3))) ) x.frstchar=c;  //scan for real char
-        if (firstwasspace==false && x.frstchar=='[' && c==32)  { if(buf(3)==']' || buf(4)==']' ){ x.frstchar=96,xword0=0; } }
-          fl = 0;
-        if ((c4&0xff) != 0) {
-            if (isalpha(c4&0xff)) fl = 1;
-            else if (ispunct(c4&0xff)) fl = 2;
-            else if (isspace(c4&0xff)) fl = 3;
-            else if ((c4&0xff) == 0xff) fl = 4;
-            else if ((c4&0xff) < 16) fl = 5;
-            else if ((c4&0xff) < 64) fl = 6;
-            else fl = 7;
-        }
-        mask = (mask<<3)|fl;
-        //0x2671756F743B
-        if (c=='(' || c=='{' || c=='[' || c=='<' || (x.wrtc4&0xffffffff)==0x266C743B/*|| x.wrtc4==0x26616D703B6C743B*/ ) { opened=c;}
-        else if (c==')' || c=='}' || c==']' || c=='>' || c==QUOTE || c==APOSTROPHE ||
-        (x.wrtc4&0xffffffff)==0x2667743B /*||
-            x.wrtc4==0x26616D703B67743B||   (x.wrtc4&0xffffffffffff)==0x2671756F743B*/) {opened=0;}
-      const bool is_newline_pC = pC==NEW_LINE || pC==0;
-        if( (c>='a' && c<='z')|| c>=128 || (c>0 && c<4))  {
-        expr0chars = expr0chars<<8|c; // last 4 consecutive letters
-        expr0=hash(expr0,c);
-        exprlen0=min(exprlen0+1,48);//maxwordlen
-      } else {
-        expr0chars=0;
-        exprlen0=0;
-        if( (c==SPACE || c==10 || c==0) && (is_letter_pC || pC==APOSTROPHE || pC==QUOTE) ) {//
-          expr4=expr3;
-          expr3=expr2;
-          expr2=expr1;
-          expr1=expr0;
-          expr0=0;
-        } else if(c==APOSTROPHE || c==QUOTE || ( (c==10 || c==0) && pC==SPACE) || (c==SPACE && is_newline_pC) ) {//
-          //ignore
-        } else {
-          expr4=expr3=expr2=expr1=expr0=0;
-        }
-       
-    }
-    }
-
-    int predict(const U8 pdf_text_parser_state,int val1=0,int val2=0) {
-      if (x.bpos==0) {
-       int i=0;
-       const U32 dist =  llog(x.blpos-wpos[word0&(wpos.size()-1)]);
-       bool istext=!(x.filetype==DEFAULT || x.filetype==DECA || x.filetype==IMGUNK || x.filetype==EXE || x.filetype==ARM);
-        x.tmask=i=((c<5)<<5)|((0)<<4)| ((dist==0)<<3)| ((/*doXML=true?1:*/0)<<2)| ((opened)<<6)| ((istemplate)<<1)| islink;
-        if (x.filetype==DEFAULT || x.filetype==DECA || x.filetype==IMGUNK) i=0;
-        else i=(hash(6,i+pdf_text_parser_state*1024)>>16);
-
- if (val2!=-1){
-        //256+ hash 513+ none
-        if ((pdf_text_parser_state!=0 || x.filetype==DECA|| x.filetype==EXE|| x.filetype==IMGUNK)){
-          if (pdf_text_parser_state!=0 ){           
-             if(exprlen0>=1) {
-             const int wl_wme_mbc=min(exprlen0,1+3)<<2;
-             cm.set(hash(++i, wl_wme_mbc, c));
-             }      else {cm.set();i++;}
-
-             if(exprlen0>=2) {
-             const int wl_wme_mbc=min(exprlen0,2+3)<<2;
-             cm.set(hash(++i, wl_wme_mbc, expr0chars&0xffff));
-             }      else {cm.set();i++;}
-
-             if(exprlen0>=3) {
-             const int wl_wme_mbc=min(exprlen0,3+3)<<2;
-             cm.set(hash(++i, wl_wme_mbc, expr0chars&0xffffff));
-             }      else {cm.set();i++;}
-
-             if(exprlen0>=4) {
-             const int wl_wme_mbc=min(exprlen0,4+3)<<2;
-             cm.set(hash(++i, wl_wme_mbc, expr0chars));
-             }   else {cm.set();i++;}
-             cm.set(hash(++i,x.spaces, (x.words&255), (numbers&255), (x.pwords&255)));//spaces...
-             for(int j=0;j<5;j++)  {cm.set();++i;}
-          } else for(int j=0;j<10;j++)  {cm.set();++i;}
-           
-        }
-        else {
-            cm.set(hash(++i,x.spafdo, x.spaces&0x7ff ,ccword));
-            cm.set(hash(++i,line0,inkeyw[w]));
-            cm.set(hash(++i,x.frstchar, c));
-            cm.set(hash(++i,x.col, x.frstchar, (lastUpper<x.col)*4+(mask2&3)));//?
-            cm.set(hash(++i,x.spaces, (x.words&255), (numbers&255), (x.pwords&255)));//spaces...
-        
-            //cm.set(x.spaces&0x7ff);
-           /* cm.set(hash(++i,word0, word2,wordGap, word6));
-            cm.set(hash(++i,word1, c,ccword));  
-            cm.set(hash(++i,word0, word1,inkeyw[w]));  
-            cm.set(hash(++i,word0, word6, lastDigit<wordGap+x.wordlen));
-            cm.set(hash(++i,word0, cword0,inkeyw[word1&(inkeyw.size()-1)]));*/
-            cm.set(hash(++i,number0, word1,wordGap, 0));
-            cm.set(hash(++i,number1, c,ccword));  
-            cm.set(hash(++i,number0, number1,inkeyw[w]));  
-            cm.set(hash(++i,word0, 0, lastDigit<wordGap+x.wordlen));
-            cm.set(hash(++i,number0, cword0,inkeyw[word1&(inkeyw.size()-1)]));
-        }
-        U32 h=x.wordcount*64+x.spacecount+numbercount*128;
-        if (x.filetype!=DECA){
-        if ( scountset==false /*&&!(x.frstchar=='{' && istemplate!=0)*/   ){
-       
-            cm.set(hash(++i,x.wordlen,wrth,x.col));
-            cm.set(hash(++i,c,x.spacecount/2,wordGap));
-           // U32 h=x.wordcount*64+x.spacecount+x.numbercount*128;
-            cm.set(hash(++i,c,h,ccword));
-            cm.set(hash(++i,x.frstchar,h,lastLetter));
-            cm.set(hash(data0,word1, number1,type0&0xFFF));//cm.set(hash(data0,word1,word2, type0&0xFFF));
-            cm.set(h);
-            cm.set(hash(++i,h,x.spafdo)); 
-            U32 d=c4&0xf0ff;
-            cm.set(hash(++i,d,x.frstchar,ccword));
-        }else {
-            for(int j=0;j<8;j++)  {cm.set();++i;}
-        }
-        }
-        h=word0*271;
-        h=h+buf(1);U32 isfword=x.filetype==DECA?-1:fword;
-        if (/*x.filetype==DEFAULT||*/x.filetype==DECA){
-        cm.set(hash(++i,h));
-        cm.set(hash(++i,word0,0)); 
-        cm.set(hash(++i,data0));// ,iCtx()
-        }else if (pdf_text_parser_state>0){
-            cm.set(hash(h,pdf_text_parser_state));
-        cm.set(hash(++i,word0)); 
-        cm.set(hash(++i,data0));//  
-            }else{
-        cm.set(hash(h));
-        cm.set(hash(++i,word0,inkeyw[w],x.tmask)); 
-        cm.set(hash(++i,data0,inkeyw[w],x.tmask));//  
-        }
-        //if (wrdhsh) cm.set(hash(++i,wrdhsh,buf(wpos[word1&(wpos.size()-1)]))); else  {cm.set();++i;}
-        cm.set(hash(++i,data0,c, word1)); 
-        cm.set(hash(++i,h, word1)); 
-        cm.set(hash(++i,word0, word1));
-        cm.set(hash(++i,h, word1,word2/*,lastUpper<x.wordlen*/));
-        cm.set(hash(++i,text0&0xffffff));
-        cm.set(text0&0xfffff);
-         
-        if (doXML==true && pdf_text_parser_state==0){
-            if ( scountset==false ){
-            cm.set(hash(++i,word0,number0,wpword1,xword0));
-            cm.set(hash(++i,word0,number0,wpword1, xword1));  //wiki 
-            cm.set(hash(++i,word0,number0,wpword1, xword2));  //
-            cm.set(hash(++i,word0,number0,wpword1, xword3));  //
-            cm.set(hash(++i,x.frstchar, xword2,inkeyw[w]));
-            }else {
-            for(int j=0;j<5;j++)  {cm.set();++i;}
-        }
-        }else{
-            //for(int j=0;j<5;j++)  {cm.set();++i;}
-            cm.set(hash(++i+512,nl1));    //chars occurring in this paragraph (order 0)
-     cm.set();++i;//cm.set(hash(++i+512,nl1,c));  //chars occurring in this paragraph (order 1)
-      cm.set(hash(++i+512,x.frstchar));   //chars occurring in a paragraph that began with frstchar (order 0)
-      //cm.set(hash(++i+512,x.frstchar,c)); //chars occurring in a paragraph that began with frstchar (order 1)*/
-      if (lastSpace<1024*4)       cm.set(hash(++i+512,h, word1, word4)); else cm.set();++i;
-      cm.set(hash(++i+512,x.col,x.wcol,nl1-nl));//cm.set();   ++i;
-    //  cm.set(hash(expr0,expr1,expr2,expr3,expr4));
-           
-       // cm.set(hash(++i,expr0,expr1,expr2));
-        }
-        
-        cm.set(hash(++i,word0, cword0,isfword));
-        if (lastSpace<1024*4){
-            if (witmode==true) cm.set(hash(++i,h, word2));else cm.set(hash(++i,h, word2,isfword));//,isfword fp.log
-            cm.set(hash(++i,h, word3));
-            cm.set(hash(++i,h, word4));
-            cm.set((istext==true)?wpword1:hash(++i,h, word5));
-            cm.set(hash(++i,h, word1,word3));
-            cm.set(hash(++i,h, word2,word3));
-        }else{
-            for(int j=0;j<6;j++)  {cm.set();++i;}
-        }
-        if  (x.filetype!=DECA ){
-            if (witmode==true  )  {
-                cm.set( (fword));++i;
-            cm.set(hash(++i,isfword,c));
-            }else{
-            
-            cm.set(hash(++i,isfword));
-            cm.set(hash(++i,isfword,c));}
-        }
-        cm.set(hash(++i,opened,c,dist!=0,pdf_text_parser_state));
-        //cm.set(hash(++i,opened,word0));
-        if (x.dictonline==true) cm.set(hash(buf(1),wchar1,wchar2)); else {cm.set();}
-        cm.set(x.f4&0x00000fff);
-        cm.set(x.f4); 
-        
-        const U8 pC_above = (x.col>(nl-nl1))?0: buf[nl1+x.col-1];
-        const bool is_new_line_start = x.col==0 && nl1>0;
-        const bool is_prev_char_match_above = buf(1)==pC_above && x.col!=0 && nl1!=0;
-        const U32 above_ctx = above<<1|U32(is_prev_char_match_above);
-        if (pdf_text_parser_state==0 &&!(x.frstchar=='{' && istemplate!=0) &&scountset==false && x.col<((U32)max(255,val2))&& x.filetype!=DECA){
-            cm.set(hash(++i,((wnl1-wnl2)<<16)|nl1-nl2,(wcol<<8)|x.col,buf(1),above_ctx));  
-            cm.set(hash(++i,buf(1),above_ctx,above^above1,wabove ));
-            cm.set(hash(++i,x.col,wcol,buf(1))); 
-            cm.set(hash(++i,x.col,wcol,c==32));  
-        } 
-        else {
-          if (x.filetype!=DECA )
-            for(int j=0;j<4;j++)  {cm.set();++i;}
-        }
- 
-        if (x.wordlen) cm.set(hash(++i, word0, dist>>4));    
-        else  {cm.set();++i;}
-        if (x.wordlen)cm.set(hash(++i,word1, data0,dist>>2));   
-        else {cm.set();++i;}
-        cm.set(hash(++i,buf(1),word0,llog(x.blpos-wpos[word2&(wpos.size()-1)])>>2));
-       
-            cm.set(hash(++i,mask)); 
-            cm.set(hash(++i,mask,buf(1))); 
-            if (pdf_text_parser_state==0&&scountset==false &&!(x.frstchar=='{' && istemplate!=0))    cm.set(hash(++i,mask&0x1ff,x.col,wcol)); else  {cm.set();++i;}
-            cm.set(hash(++i,mask,c4&0x00ffff00)); 
-            cm.set(hash(++i,mask&0x1ff,x.f4&0x00fff0)); 
-            cm.set(hash(++i,h, llog(wordGap), mask&0x1FF, ///// dis>30
-             ((istemplate)<<9)|  
-             ((islink)<<8)|
-             ((linespace > 4)<<7)|
-             ((x.wordlen1 > 3)<<6)|
-             ((x.wordlen > 0)<<5)|
-             ((x.spafdo == x.wordlen + 2)<<4)|
-             ((x.spafdo == x.wordlen + x.wordlen1 + 3)<<3)|
-             ((x.spafdo >= lastLetter + x.wordlen1 + wordGap)<<2)|
-             ((lastUpper < lastLetter + x.wordlen1)<<1)|
-             (lastUpper < x.wordlen + x.wordlen1 + wordGap)
-             ));
-      
-        
-        if (x.wordlen1 && pdf_text_parser_state==0&&scountset==false &&!(x.frstchar=='{' && istemplate!=0))            cm.set(hash((wcol<<8)|x.col,x.wordlen1,above,above1,c4&0xfF)); else cm.set(); //wordlist 
-        if (wrdhsh)          cm.set(hash(++i,mask2&0x3F, wrdhsh&0xFFF, (0x100|firstLetter)*(x.wordlen<6),(wordGap>4)*2+(x.wordlen1>5)) ); else i++,cm.set();//?
-        hq=hash((wcol<<8)|x.col,above^above1,numbers&127,x.filetype==DECA?c4&0xfF:llog(x.blpos-wpos[word0&(wpos.size()-1)]));
-        if (  pdf_text_parser_state==0&&scountset==false &&!(x.frstchar=='{' && istemplate!=0)) cm.set(hash(++i,(wcol<<8)|x.col,above^above1,above2 , ((islink)<<8)|
-             ((linespace > 4)<<7)|
-             ((x.wordlen1 > 3)<<6)|
-             ((x.wordlen > 0)<<5)|
-             ((x.spafdo == x.wordlen + 2)<<4)|
-             ((x.spafdo == x.wordlen + x.wordlen1 + 3)<<3)|
-             ((x.spafdo >= lastLetter + x.wordlen1 + wordGap)<<2)|
-             ((lastUpper < lastLetter + x.wordlen1)<<1)|
-             (lastUpper < x.wordlen + x.wordlen1 + wordGap)));   else  {cm.set();++i;}
-        cm.set(hash((*pWord).Hash[2], h));
-        
- }
-    }
-    if (val2==-1) return 1;
-    
-    if(x.bpos==0) {
-
-    U32 g=c4&0xff; // group identifier
-         if('0'<=g && g<='9') g='0'; //all digits are in one group
-    else if('A'<=g && g<='Z') g='A'; //all uppercase letters are in one group
-    else if('a'<=g && g<='z' || x.wstat==true) g='a'; //all lowercase letters are in one group
-    else if(g>=128) g=128;
-
-    const bool to_be_collapsed = (g=='0' || g=='A' || g=='a') && g == (g_ascii_lo&0xff);
-    if(!to_be_collapsed) {
-      g_ascii_hi <<= 8;
-      g_ascii_hi  |= g_ascii_lo>>(64-8);
-      g_ascii_lo <<= 8;
-      g_ascii_lo  |= g;
-    }
- if( linespace>0 || x.dictonline==true) {
-    U64 i = to_be_collapsed*8;//i=i*65;
-    U32 g_a_hi=(g_ascii_lo>>32),g_a_lo=g_ascii_lo&0x00000000ffffffff;
-    cm.set(hash( (++i), g_a_hi, g_a_lo, g_ascii_hi&0x00000000ffffffff ));  // last 12 groups
-    cm.set(hash( (++i), g_a_hi, g_a_lo));                                  // last 8 groups
-    cm.set(hash( (++i), g_a_hi&0x0000ffff, g_a_lo ));                      // last 6 groups
-    cm.set(hash( (++i), g_a_lo));                                          // last 4 groups
-    cm.set(hash( (++i), g_a_lo&0x0000ffff));                               // last 2 groups
-    cm.set(hash( (++i), g_a_hi&0x00ffffff, g_a_lo , c4&0x0000ffff ));    // last 7 groups + last 2 chars
-    cm.set(hash( (++i), g_a_hi&0x000000ff, g_a_lo , c4&0x00ffffff ));    // last 5 groups + last 3 chars
- } else for(int i=0;i<7;i++) cm.set();
-  }
-     return hq;
-    }
-  };
-  
-  U8 pdf_text_parser_state,math_state,pre_state; // 0,1,2,3
-  Info info_normal;
-  Info info_pdf;
-  Info math;
-  Info pre;
-  Info xhtml;
-  U32 hq;
-public:
-  wordModel1( BlockData& bd,U32 val=16): x(bd),buf(bd.buf),N(64+7),cm(CMlimit(MEM()*val), N,M_WORD),cm1(CMlimit(MEM()), 1,M_WORD),pdf_text_parser_state(0),math_state(0),pre_state(0),
-  info_normal(bd,0,cm), info_pdf(bd,0,cm), math(bd,0,cm),pre(bd,0,cm),xhtml(bd,0,cm),hq(0){
-  
-   }
-   int inputs() {return N*cm.inputs()+7;}
-   int nets() {return 0;}
-  int netcount() {return 0;}
-
-  #define bswap64(n) ((bswap((n&0xFFFFFFFF00000000)>>32))|((bswap(n&0x00000000FFFFFFFF))<<32))
-   int p(Mixer& m,int val1=0,int val2=0)  {
-    if (x.bpos==0) {
-      const U8 c1=x.c4;
-      //extract text from pdf
-      if((x.wrtc4&0xffffffff)==0x0A62740A&& x.wrtstatus==0 && x.wrtLoaded==true||x.c4==0x0a42540a) {//  "\nBT\n"  "\nbt\n" 
-          pdf_text_parser_state=1;x.inpdf=true;
-      } // Begin Text
-      else if ((x.wrtc4&0xffffffff)==0x0A65740A && x.wrtstatus==0 && x.wrtLoaded==true||x.c4==0x0a45540a) {// "\nET\n" "\net\n" 
-          pdf_text_parser_state=0;x.inpdf=false;
-      } // End Text
-      bool do_pdf_process=true;
-      if(pdf_text_parser_state!=0) {
-        const U8 pC=x.c4>>8;
-        if(pC!='\\') {
-             if(c1=='[') {pdf_text_parser_state|=2;} //array begins
-        else if(c1==']') {pdf_text_parser_state&=(255-2);}
-        else if(c1=='(') {pdf_text_parser_state|=4; do_pdf_process=false;} //signal: start text extraction
-        else if(c1==')') {pdf_text_parser_state&=(255-4);} //signal: start pdf gap processing
-        }
-      }
-      // ':&<math& >' '|&<math& >' ' &<math& >' '\n&<math& >'
-      if(x.wrtc4==0x3A263C6D61746826|| x.wrtc4==0x20263C6D61746826||x.wrtc4==0x10263C6D61746826||/*(x.wrtc4&0xffffffff)==0x6D617468||*/ x.wrtc4==0x7C263C6D61746826 ) {//  
-          math_state=2;         
-      } // Begin math
-      else if (math_state &&(x.wrtc4&0xff)==0x26  ||(x.wrtc4&0xff)==0x20) {//  
-          math_state=0; 
-      } // End math
-      //else if (math_state==1 && ((x.c4>>8)&0xff)=='>') {//  
-      //    math_state=2; 
-      //} // End math
-      //&<pre&>
-      if((x.wrtc4&0xffffffffffffff)==0x263C707265263E ) {//  
-          pre_state=4; 
-      } // Begin pre
-      else if (pre_state &&(x.wrtc4&0xffffff)==0x263C2F  ) {// '</p' 
-          pre_state=0; 
-      } // End pre
-      const bool is_pdftext = (pdf_text_parser_state&4)!=0;
-      if(is_pdftext) {
-        if(do_pdf_process) {
-          //printf("%c ",c1);  //debug: print the extracted pdf text
-          const bool is_extended_char =x.wrtstatus==-1 || c1>=128|| (c1 <4 && c1>0 && x.wrtLoaded==true && x.wrtstatus==0);
-          info_pdf.process_char(x.wrtstatus,val1,val2);
-        }
-        hq=info_pdf.predict(pdf_text_parser_state,val1,val2);
-      } else   if(math_state==2) {
-          math.process_char(x.wrtstatus,val1,val2);
-          hq=math.predict(math_state,val1,val2);
-      } else   if(pre_state==4) {
-          pre.process_char(x.wrtstatus,val1,val2);
-          hq=pre.predict(pre_state,val1,val2);
-      } else   if( x.istex==false && x.ishtml==true) {
-          xhtml.process_char(x.wrtstatus,val1,val2);
-          hq=xhtml.predict(0,val1,val2);
-      } else {
-        const bool is_textblock = (x.filetype==TEXT||x.filetype==TEXT0|| x.filetype==TXTUTF8||x.filetype==EOLTEXT||x.filetype==DICTTXT||x.filetype==BIGTEXT||x.filetype==NOWRT);
-        const bool is_extended_char = is_textblock && c1>=128;
-        info_normal.process_char(x.wrtstatus,val1,val2);
-        hq=info_normal.predict(pdf_text_parser_state,val1,val2);
-      }      
-      cm1.set(x.w4);
-    }
-    cm.mix(m);
-    cm1.mix(m);
-    return hq;
-  }
-  
- virtual ~wordModel1(){ }
-};
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
 
@@ -3088,11 +923,6 @@ int p(Mixer& m,int val1=0,int val2=0){
 }
   virtual ~ppmdModel1(){ }
 };
-
-
-
-
-
 
 
 /*
