+++ prt/helper.hpp
@@ -28,13 +28,15 @@ constexpr auto isPowerOf2(T x) -> bool {
 +    (((x) & 0x00ff0000) >>  8) | \
 +    (((x) & 0x0000ff00) <<  8) | \
 +    (((x) & 0x000000ff) << 24))
-
+#define bswap64(n) ((bswap((n&0xFFFFFFFF00000000)>>32))|((bswap(n&0x00000000FFFFFFFF))<<32))
 U8 Clip(int const Px);
 
 #define TAB 0x09
 #define NEW_LINE 0x0A
 #define CARRIAGE_RETURN 0x0D
 #define SPACE 0x20
+#define QUOTE 0x22
+#define APOSTROPHE 0x27
 
  int min(int a, int b);
  int max(int a, int b);
@@ -44,3 +46,5 @@ U8 Clamp4(const int Px, const U8 n1, const U8 n2, const U8 n3, const U8 n4);
   U32 LogQt(const U8 Px, const U8 bits);
 
   U8 Paeth(U8 W, U8 N, U8 NW);
+  #define MAX_WORD_SIZE 64
+   bool CharInArray(const char c, const char a[], const int len) ;
