+++ prt/stemmer/stemmer.hpp
@@ -1,0 +1,1453 @@
+#pragma once
+
+
+class Word {
+public:
+  U8 Letters[MAX_WORD_SIZE];
+  U8 Start, End;
+  U32 Hash[4], Type, Language;
+  Word() : Start(0), End(0),  Type(0), Language(0) {
+    memset(&Letters[0], 0, sizeof(U8)*MAX_WORD_SIZE);
+    memset(&Hash[0], 0, sizeof(U32)*4);
+  }
+  bool operator==(const char *s) const {
+    size_t len=strlen(s);
+    return ((size_t)(End-Start+(Letters[Start]!=0))==len && memcmp(&Letters[Start], s, len)==0);
+  }
+  bool operator!=(const char *s) const {
+    return !operator==(s);
+  }
+  void operator+=(const char c) {
+    if (End<MAX_WORD_SIZE-1){
+      End+=(Letters[End]>0);
+      Letters[End]=tolower(c);
+    }
+  }
+  U8 operator[](U8 i) const {
+    return (End-Start>=i)?Letters[Start+i]:0;
+  }
+  U8 operator()(U8 i) const {
+    return (End-Start>=i)?Letters[End-i]:0;
+  }
+  U32 Length() const {
+    if (Letters[Start]!=0)
+      return End-Start+1;
+    return 0;
+  }
+  void GetHashes() {
+    Hash[0] = 0xc01dflu, Hash[1] = ~Hash[0];
+    for (int i=Start; i<=End; i++) {
+      U8 l = Letters[i];
+      Hash[0]^=hash(Hash[0], l, i);
+      Hash[1]^=hash(Hash[1], 
+        ((l&0x80)==0)?l&0x5F:
+        ((l&0xC0)==0x80)?l&0x3F:
+        ((l&0xE0)==0xC0)?l&0x1F:
+        ((l&0xF0)==0xE0)?l&0xF:l&0x7
+      );
+    }
+    Hash[2] = (~Hash[0])^Hash[1];
+    Hash[3] = (~Hash[1])^Hash[0];
+  }
+  bool ChangeSuffix(const char *OldSuffix, const char *NewSuffix) {
+    size_t len=strlen(OldSuffix);
+    if (Length()>len && memcmp(&Letters[End-len+1], OldSuffix, len)==0){
+      size_t n=strlen(NewSuffix);
+      if (n>0){
+        memcpy(&Letters[End-int(len)+1], NewSuffix, min(MAX_WORD_SIZE-1,End+int(n))-End);
+        End=min(MAX_WORD_SIZE-1, End-int(len)+int(n));
+      }
+      else
+        End-=U8(len);
+      return true;
+    }
+    return false;
+  }
+  bool MatchesAny(const char* a[], const int count) {
+    int i=0;
+    size_t len = (size_t)Length();
+    for (; i<count && (len!=strlen(a[i]) || memcmp(&Letters[Start], a[i], len)!=0); i++);
+    return i<count;
+  }
+  bool EndsWith(const char *Suffix) const {
+    size_t len=strlen(Suffix);
+    return (Length()>len && memcmp(&Letters[End-len+1], Suffix, len)==0);
+  }
+  bool StartsWith(const char *Prefix) const {
+    size_t len=strlen(Prefix);
+    return (Length()>len && memcmp(&Letters[Start], Prefix, len)==0);
+  }
+  void print() const {
+    for(int r=Start;r<=End;r++)
+      printf("%c",(char)Letters[r]);
+    printf("\n");
+  }
+};
+
+class Segment1 {
+public:
+  Word FirstWord; // useful following questions
+  U32 WordCount;
+  U32 NumCount;
+};
+
+class Sentence : public Segment1 {
+public:
+  enum Types { // possible sentence types, excluding Imperative
+    Declarative,
+    Interrogative,
+    Exclamative,
+    Count
+  };
+  Types Type;
+  U32 SegmentCount;
+  U32 VerbIndex; // relative position of last detected verb
+  U32 NounIndex; // relative position of last detected noun
+  U32 CapitalIndex; // relative position of last capitalized word, excluding the initial word of this sentence
+  Word lastVerb, lastNoun, lastCapital;
+};
+
+class Paragraph {
+public:
+  U32 SentenceCount, TypeCount[Sentence::Types::Count], TypeMask;
+};
+
+class Language {
+public:
+   enum Flags {
+    Verb                   = (1<<0),
+    Noun                   = (1<<1)
+  };
+  enum Ids {
+    Unknown,
+    English,
+    French,
+    German,
+    Count
+  };
+  virtual ~Language() {};
+  virtual bool IsAbbreviation(Word *W) = 0;
+};
+
+class English: public Language {
+private:
+  static const int NUM_ABBREV = 6;
+  const char *Abbreviations[NUM_ABBREV]={ "mr","mrs","ms","dr","st","jr" };
+public:
+  enum Flags {
+   
+    Adjective              = (1<<2),
+    Plural                 = (1<<3),
+    Male                   = (1<<4),
+    Female                 = (1<<5),
+    Negation               = (1<<6),
+    PastTense              = (1<<7)|Verb,
+    PresentParticiple      = (1<<8)|Verb,
+    AdjectiveSuperlative   = (1<<9)|Adjective,
+    AdjectiveWithout       = (1<<10)|Adjective,
+    AdjectiveFull          = (1<<11)|Adjective,
+    AdverbOfManner         = (1<<12),
+    SuffixNESS             = (1<<13),
+    SuffixITY              = (1<<14)|Noun,
+    SuffixCapable          = (1<<15),
+    SuffixNCE              = (1<<16),
+    SuffixNT               = (1<<17),
+    SuffixION              = (1<<18),
+    SuffixAL               = (1<<19)|Adjective,
+    SuffixIC               = (1<<20)|Adjective,
+    SuffixIVE              = (1<<21),
+    SuffixOUS              = (1<<22)|Adjective,
+    PrefixOver             = (1<<23),
+    PrefixUnder            = (1<<24)
+  };
+  bool IsAbbreviation(Word *W) { return W->MatchesAny(Abbreviations, NUM_ABBREV); };
+};
+
+class French: public Language {
+private:
+  static const int NUM_ABBREV = 2;
+  const char *Abbreviations[NUM_ABBREV]={ "m","mm" };
+public:
+  enum Flags {
+    
+    Adjective              = (1<<2),
+    Plural                 = (1<<3)
+  };
+  bool IsAbbreviation(Word *W) { return W->MatchesAny(Abbreviations, NUM_ABBREV); };
+};
+
+class German : public Language {
+private:
+  static const int NUM_ABBREV = 3;
+  const char *Abbreviations[NUM_ABBREV]={ "fr","hr","hrn" };
+public:
+  enum Flags {
+    Adjective              = (1<<2),
+    Plural                 = (1<<3),
+    Female                 = (1<<4)
+  };
+  bool IsAbbreviation(Word *W) { return W->MatchesAny(Abbreviations, NUM_ABBREV); };
+};
+//////////////////////////// Stemming routines /////////////////////////
+
+class Stemmer {
+    protected:
+  U32 GetRegion(const Word *W, const U32 From) {
+    bool hasVowel = false;
+    for (int i=W->Start+From; i<=W->End; i++) {
+      if (IsVowel(W->Letters[i])) {
+        hasVowel = true;
+        continue;
+      }
+      else if (hasVowel)
+        return i-W->Start+1;
+    }
+    return W->Start+W->Length();
+  }
+  bool SuffixInRn(const Word *W, const U32 Rn, const char *Suffix) {
+    return (W->Start!=W->End && Rn<=W->Length()-strlen(Suffix));
+  }
+public:
+  virtual ~Stemmer() {};
+  virtual bool IsVowel(const char c) = 0;
+  virtual void Hash(Word *W) = 0;
+  virtual bool Stem(Word *W) = 0;
+};
+
+
+class EnglishStemmer: public Stemmer {
+private:
+  static const int NUM_VOWELS = 6;
+  const char Vowels[NUM_VOWELS]={'a','e','i','o','u','y'};
+  static const int NUM_DOUBLES = 9;
+  const char Doubles[NUM_DOUBLES]={'b','d','f','g','m','n','p','r','t'};
+  static const int NUM_LI_ENDINGS = 10;
+  const char LiEndings[NUM_LI_ENDINGS]={'c','d','e','g','h','k','m','n','r','t'};
+  static const int NUM_NON_SHORT_CONSONANTS = 3;
+  const char NonShortConsonants[NUM_NON_SHORT_CONSONANTS]={'w','x','Y'};
+  static const int NUM_MALE_WORDS = 9;
+  const char *MaleWords[NUM_MALE_WORDS]={"he","him","his","himself","man","men","boy","husband","actor"};
+  static const int NUM_FEMALE_WORDS = 8;
+  const char *FemaleWords[NUM_FEMALE_WORDS]={"she","her","herself","woman","women","girl","wife","actress"};
+  static const int NUM_COMMON_WORDS = 12;
+  const char *CommonWords[NUM_COMMON_WORDS]={"the","be","to","of","and","in","that","you","have","with","from","but"};
+  static const int NUM_SUFFIXES_STEP0 = 3;
+  const char *SuffixesStep0[NUM_SUFFIXES_STEP0]={"'s'","'s","'"};
+  static const int NUM_SUFFIXES_STEP1b = 6;
+  const char *SuffixesStep1b[NUM_SUFFIXES_STEP1b]={"eedly","eed","ed","edly","ing","ingly"};
+  const U32 TypesStep1b[NUM_SUFFIXES_STEP1b]={English::AdverbOfManner,0,English::PastTense,English::AdverbOfManner|English::PastTense,English::PresentParticiple,English::AdverbOfManner|English::PresentParticiple};
+  static const int NUM_SUFFIXES_STEP2 = 22;
+  const char *(SuffixesStep2[NUM_SUFFIXES_STEP2])[2]={
+    {"ization", "ize"},
+    {"ational", "ate"},
+    {"ousness", "ous"},
+    {"iveness", "ive"},
+    {"fulness", "ful"},
+    {"tional", "tion"},
+    {"lessli", "less"},
+    {"biliti", "ble"},
+    {"entli", "ent"},
+    {"ation", "ate"},
+    {"alism", "al"},
+    {"aliti", "al"},
+    {"fulli", "ful"},
+    {"ousli", "ous"},
+    {"iviti", "ive"},
+    {"enci", "ence"},
+    {"anci", "ance"},
+    {"abli", "able"},
+    {"izer", "ize"},
+    {"ator", "ate"},
+    {"alli", "al"},
+    {"bli", "ble"}
+  };
+  const U32 TypesStep2[NUM_SUFFIXES_STEP2]={
+    English::SuffixION,
+    English::SuffixION|English::SuffixAL,
+    English::SuffixNESS,
+    English::SuffixNESS,
+    English::SuffixNESS,
+    English::SuffixION|English::SuffixAL,
+    English::AdverbOfManner,
+    English::AdverbOfManner|English::SuffixITY,
+    English::AdverbOfManner,
+    English::SuffixION,
+    0,
+    English::SuffixITY,
+    English::AdverbOfManner,
+    English::AdverbOfManner,
+    English::SuffixITY,
+    0,
+    0,
+    English::AdverbOfManner,
+    0,
+    0,
+    English::AdverbOfManner,
+    English::AdverbOfManner
+  };
+  static const int NUM_SUFFIXES_STEP3 = 8;
+  const char *(SuffixesStep3[NUM_SUFFIXES_STEP3])[2]={
+    {"ational", "ate"},
+    {"tional", "tion"},
+    {"alize", "al"},
+    {"icate", "ic"},
+    {"iciti", "ic"},
+    {"ical", "ic"},
+    {"ful", ""},
+    {"ness", ""}
+  };
+  const U32 TypesStep3[NUM_SUFFIXES_STEP3]={English::SuffixION|English::SuffixAL,English::SuffixION|English::SuffixAL,0,0,English::SuffixITY,English::SuffixAL,English::AdjectiveFull,English::SuffixNESS};
+  static const int NUM_SUFFIXES_STEP4 = 20;
+  const char *SuffixesStep4[NUM_SUFFIXES_STEP4]={"al","ance","ence","er","ic","able","ible","ant","ement","ment","ent","ou","ism","ate","iti","ous","ive","ize","sion","tion"};
+  const U32 TypesStep4[NUM_SUFFIXES_STEP4]={
+    English::SuffixAL,
+    English::SuffixNCE,
+    English::SuffixNCE,
+    0,
+    English::SuffixIC,
+    English::SuffixCapable,
+    English::SuffixCapable,
+    English::SuffixNT,
+    0,
+    0,
+    English::SuffixNT,
+    0,
+    0,
+    0,
+    English::SuffixITY,
+    English::SuffixOUS,
+    English::SuffixIVE,
+    0,
+    English::SuffixION,
+    English::SuffixION
+  };
+  static const int NUM_EXCEPTION_REGION1 = 3;
+  const char *ExceptionsRegion1[NUM_EXCEPTION_REGION1]={"gener","arsen","commun"};
+  static const int NUM_EXCEPTIONS1 = 19;
+  const char *(Exceptions1[NUM_EXCEPTIONS1])[2]={
+    {"skis", "ski"},
+    {"skies", "sky"},
+    {"dying", "die"},
+    {"lying", "lie"},
+    {"tying", "tie"},
+    {"idly", "idl"},
+    {"gently", "gentl"},
+    {"ugly", "ugli"},
+    {"early", "earli"},
+    {"only", "onli"},
+    {"singly", "singl"},
+    {"sky", "sky"},
+    {"news", "news"},
+    {"howe", "howe"},
+    {"atlas", "atlas"},
+    {"cosmos", "cosmos"},
+    {"bias", "bias"},
+    {"andes", "andes"},
+    {"texas", "texas"}
+  };
+  const U32 TypesExceptions1[NUM_EXCEPTIONS1]={
+    English::Noun|English::Plural,
+    English::Noun|English::Plural,
+    English::PresentParticiple,
+    English::PresentParticiple,
+    English::PresentParticiple,
+    English::AdverbOfManner,
+    English::AdverbOfManner,
+    English::Adjective,
+    English::Adjective|English::AdverbOfManner,
+    0,
+    English::AdverbOfManner,
+    English::Noun,
+    English::Noun,
+    0,
+    English::Noun,
+    English::Noun,
+    English::Noun,
+    English::Noun|English::Plural,
+    English::Noun
+  };
+  static const int NUM_EXCEPTIONS2 = 8;
+  const char *Exceptions2[NUM_EXCEPTIONS2]={"inning","outing","canning","herring","earring","proceed","exceed","succeed"};
+  const U32 TypesExceptions2[NUM_EXCEPTIONS2]={English::Noun,English::Noun,English::Noun,English::Noun,English::Noun,English::Verb,English::Verb,English::Verb}; 
+  inline bool IsConsonant(const char c){
+    return !IsVowel(c);
+  }
+  inline bool IsShortConsonant(const char c){
+    return !CharInArray(c, NonShortConsonants, NUM_NON_SHORT_CONSONANTS);
+  }
+  inline bool IsDouble(const char c){
+    return CharInArray(c, Doubles, NUM_DOUBLES);
+  }
+  inline bool IsLiEnding(const char c){
+    return CharInArray(c, LiEndings, NUM_LI_ENDINGS);
+  }
+   
+  U32 GetRegion1(const Word *W){
+    for (int i=0; i<NUM_EXCEPTION_REGION1; i++){
+      if (W->StartsWith(ExceptionsRegion1[i]))
+        return U32(strlen(ExceptionsRegion1[i]));
+    }
+    return GetRegion(W, 0);
+  }
+   
+  bool EndsInShortSyllable(const Word *W){
+    if (W->End==W->Start)
+      return false;
+    else if (W->End==W->Start+1)
+      return IsVowel((*W)(1)) && IsConsonant((*W)(0));
+    else
+      return (IsConsonant((*W)(2)) && IsVowel((*W)(1)) && IsConsonant((*W)(0)) && IsShortConsonant((*W)(0)));
+  }
+  bool IsShortWord(const Word *W){
+    return (EndsInShortSyllable(W) && GetRegion1(W)==W->Length());
+  }
+  inline bool HasVowels(const Word *W){
+    for (int i=W->Start; i<=W->End; i++){
+      if (IsVowel(W->Letters[i]))
+        return true;
+    }
+    return false;
+  }
+  bool TrimStartingApostrophe(Word *W){
+    bool result=false;
+    //trim all apostrophes from the beginning
+    int cnt=0;
+    while(W->Start!=W->End && (*W)[0]=='\'') {
+      result=true;
+      W->Start++;
+      cnt++;
+    }
+    //trim the same number of apostrophes from the end (if there are)
+    while(W->Start!=W->End && (*W)(0)=='\'') {
+      if(cnt==0)break;
+      W->End--;
+      cnt--;
+    }
+    return result;
+  }
+
+  void MarkYsAsConsonants(Word *W){
+    if ((*W)[0]=='y')
+      W->Letters[W->Start]='Y';
+    for (int i=W->Start+1; i<=W->End; i++){
+      if (IsVowel(W->Letters[i-1]) && W->Letters[i]=='y')
+        W->Letters[i]='Y';
+    }
+  }
+  bool ProcessPrefixes(Word *W){
+    if (W->StartsWith("irr") && W->Length()>5 && ((*W)[3]=='a' || (*W)[3]=='e'))
+      W->Start+=2, W->Type|=English::Negation;
+    else if (W->StartsWith("over") && W->Length()>5)
+      W->Start+=4, W->Type|=English::PrefixOver;
+    else if (W->StartsWith("under") && W->Length()>6)
+      W->Start+=5, W->Type|=English::PrefixUnder;
+    else if (W->StartsWith("unn") && W->Length()>5)
+      W->Start+=2, W->Type|=English::Negation;
+    else if (W->StartsWith("non") && W->Length()>(U32)(5+((*W)[3]=='-')))
+      W->Start+=2+((*W)[3]=='-'), W->Type|=English::Negation;
+    else
+      return false;
+    return true;
+  }
+  bool ProcessSuperlatives(Word *W){
+    if (W->EndsWith("est") && W->Length()>4){
+      U8 i=W->End;
+      W->End-=3;
+      W->Type|=English::AdjectiveSuperlative;
+
+      if ((*W)(0)==(*W)(1) && (*W)(0)!='r' && !(W->Length()>=4 && memcmp("sugg",&W->Letters[W->End-3],4)==0)){
+        W->End-= ( ((*W)(0)!='f' && (*W)(0)!='l' && (*W)(0)!='s') ||
+                   (W->Length()>4 && (*W)(1)=='l' && ((*W)(2)=='u' || (*W)(3)=='u' || (*W)(3)=='v'))) &&
+                   (!(W->Length()==3 && (*W)(1)=='d' && (*W)(2)=='o'));
+        if (W->Length()==2 && ((*W)[0]!='i' || (*W)[1]!='n'))
+          W->End = i, W->Type&=~English::AdjectiveSuperlative;
+      }
+      else{
+        switch((*W)(0)){
+          case 'd': case 'k': case 'm': case 'y': break;
+          case 'g': {
+            if (!( W->Length()>3 && ((*W)(1)=='n' || (*W)(1)=='r') && memcmp("cong",&W->Letters[W->End-3],4)!=0 ))
+              W->End = i, W->Type&=~English::AdjectiveSuperlative;
+            else
+              W->End+=((*W)(2)=='a');
+            break;
+          }
+          case 'i': {W->Letters[W->End]='y'; break;}
+          case 'l': {
+            if (W->End==W->Start+1 || memcmp("mo",&W->Letters[W->End-2],2)==0)
+              W->End = i, W->Type&=~English::AdjectiveSuperlative;
+            else
+              W->End+=IsConsonant((*W)(1));
+            break;
+          }
+          case 'n': {
+            if (W->Length()<3 || IsConsonant((*W)(1)) || IsConsonant((*W)(2)))
+              W->End = i, W->Type&=~English::AdjectiveSuperlative;
+            break;
+          }
+          case 'r': {
+            if (W->Length()>3 && IsVowel((*W)(1)) && IsVowel((*W)(2)))
+              W->End+=((*W)(2)=='u') && ((*W)(1)=='a' || (*W)(1)=='i');
+            else
+              W->End = i, W->Type&=~English::AdjectiveSuperlative;
+            break;
+          }
+          case 's': {W->End++; break;}
+          case 'w': {
+            if (!(W->Length()>2 && IsVowel((*W)(1))))
+              W->End = i, W->Type&=~English::AdjectiveSuperlative;
+            break;
+          }
+          case 'h': {
+            if (!(W->Length()>2 && IsConsonant((*W)(1))))
+              W->End = i, W->Type&=~English::AdjectiveSuperlative;
+            break;
+          }
+          default: {
+            W->End+=3;
+            W->Type&=~English::AdjectiveSuperlative;
+          }
+        }
+      }
+    }
+    return (W->Type&English::AdjectiveSuperlative)>0;
+  }
+  bool Step0(Word *W){
+    for (int i=0; i<NUM_SUFFIXES_STEP0; i++){
+      if (W->EndsWith(SuffixesStep0[i])){
+        W->End-=U8(strlen(SuffixesStep0[i]));
+        W->Type|=English::Plural;
+        return true;
+      }
+    }
+    return false;
+  }
+  bool Step1a(Word *W){
+    if (W->EndsWith("sses")){
+      W->End-=2;
+      W->Type|=English::Plural;
+      return true;
+    }
+    if (W->EndsWith("ied") || W->EndsWith("ies")){
+      W->Type|=((*W)(0)=='d')?English::PastTense:English::Plural;
+      W->End-= W->Length()>4 ? 2 : 1;
+      return true;
+    }
+    if (W->EndsWith("us") || W->EndsWith("ss"))
+      return false;
+    if ((*W)(0)=='s' && W->Length()>2){
+      for (int i=W->Start;i<=W->End-2;i++){
+        if (IsVowel(W->Letters[i])){
+          W->End--;
+          W->Type|=English::Plural;
+          return true;
+        }
+      }
+    }
+    if (W->EndsWith("n't") && W->Length()>4){
+      switch ((*W)(3)){
+        case 'a': {
+          if ((*W)(4)=='c')
+            W->End-=2;
+          else
+            W->ChangeSuffix("n't","ll");
+          break;
+        }
+        case 'i': {W->ChangeSuffix("in't","m"); break;}
+        case 'o': {
+          if ((*W)(4)=='w')
+            W->ChangeSuffix("on't","ill");
+          else
+            W->End-=3;
+          break;
+        }
+        default: W->End-=3;
+      }
+      W->Type|=English::Negation;
+      return true;
+    }
+    if (W->EndsWith("hood") && W->Length()>7){
+      W->End-=4;
+      return true;
+    }
+    return false;
+  }
+  bool Step1b(Word *W, const U32 R1){
+    for (int i=0; i<NUM_SUFFIXES_STEP1b; i++){
+      if (W->EndsWith(SuffixesStep1b[i])){
+        switch(i){
+          case 0: case 1: {
+            if (SuffixInRn(W, R1, SuffixesStep1b[i]))
+              W->End-=1+i*2;
+            break;
+          }
+          default: {
+            U8 j=W->End;
+            W->End-=U8(strlen(SuffixesStep1b[i]));
+            if (HasVowels(W)){
+              if (W->EndsWith("at") || W->EndsWith("bl") || W->EndsWith("iz") || IsShortWord(W))
+                (*W)+='e';
+              else if (W->Length()>2){
+                if ((*W)(0)==(*W)(1) && IsDouble((*W)(0)))
+                  W->End--;
+                else if (i==2 || i==3){
+                  switch((*W)(0)){
+                    case 'c': case 's': case 'v': {W->End+=!(W->EndsWith("ss") || W->EndsWith("ias")); break;}
+                    case 'd': {
+                      static const char nAllowed[4] = {'a','e','i','o'};
+                      W->End+=IsVowel((*W)(1)) && (!CharInArray((*W)(2), nAllowed, 4)); break;
+                    }
+                    case 'k': {W->End+=W->EndsWith("uak"); break;}
+                    case 'l': {
+                      static const char Allowed1[10] = {'b','c','d','f','g','k','p','t','y','z'};
+                      static const char Allowed2[4] = {'a','i','o','u'};
+                      W->End+= CharInArray((*W)(1), Allowed1, 10) ||
+                                (CharInArray((*W)(1), Allowed2, 4) && IsConsonant((*W)(2)));
+                      break;
+                    }
+                  }
+                }
+                else if (i>=4){
+                  switch((*W)(0)){
+                    case 'd': {
+                      if (IsVowel((*W)(1)) && (*W)(2)!='a' && (*W)(2)!='e' && (*W)(2)!='o')
+                        (*W)+='e';
+                      break;
+                    }
+                    case 'g': {
+                      static const char Allowed[7] = {'a','d','e','i','l','r','u'};
+                      if (
+                        CharInArray((*W)(1), Allowed, 7) || (
+                         (*W)(1)=='n' && (
+                          (*W)(2)=='e' ||
+                          ((*W)(2)=='u' && (*W)(3)!='b' && (*W)(3)!='d') ||
+                          ((*W)(2)=='a' && ((*W)(3)=='r' || ((*W)(3)=='h' && (*W)(4)=='c'))) ||
+                          (W->EndsWith("ring") && ((*W)(4)=='c' || (*W)(4)=='f'))
+                         )
+                        ) 
+                      )
+                        (*W)+='e';
+                      break;
+                    }
+                    case 'l': {
+                      if (!((*W)(1)=='l' || (*W)(1)=='r' || (*W)(1)=='w' || (IsVowel((*W)(1)) && IsVowel((*W)(2)))))
+                        (*W)+='e';
+                      if (W->EndsWith("uell") && W->Length()>4 && (*W)(4)!='q')
+                        W->End--;
+                      break;
+                    }
+                    case 'r': {
+                      if ((
+                        ((*W)(1)=='i' && (*W)(2)!='a' && (*W)(2)!='e' && (*W)(2)!='o') ||
+                        ((*W)(1)=='a' && (!((*W)(2)=='e' || (*W)(2)=='o' || ((*W)(2)=='l' && (*W)(3)=='l')))) ||
+                        ((*W)(1)=='o' && (!((*W)(2)=='o' || ((*W)(2)=='t' && (*W)(3)!='s')))) ||
+                        (*W)(1)=='c' || (*W)(1)=='t') && (!W->EndsWith("str"))
+                      )
+                        (*W)+='e';
+                      break;
+                    }
+                    case 't': {
+                      if ((*W)(1)=='o' && (*W)(2)!='g' && (*W)(2)!='l' && (*W)(2)!='i' && (*W)(2)!='o')
+                        (*W)+='e';
+                      break;
+                    }
+                    case 'u': {
+                      if (!(W->Length()>3 && IsVowel((*W)(1)) && IsVowel((*W)(2))))
+                        (*W)+='e';
+                      break;
+                    }
+                    case 'z': {
+                      if (W->EndsWith("izz") && W->Length()>3 && ((*W)(3)=='h' || (*W)(3)=='u'))
+                        W->End--;
+                      else if ((*W)(1)!='t' && (*W)(1)!='z')
+                        (*W)+='e';
+                      break;
+                    }
+                    case 'k': {
+                      if (W->EndsWith("uak"))
+                        (*W)+='e';
+                      break;
+                    }
+                    case 'b': case 'c': case 's': case 'v': {
+                      if (!(
+                        ((*W)(0)=='b' && ((*W)(1)=='m' || (*W)(1)=='r')) ||
+                        W->EndsWith("ss") || W->EndsWith("ias") || (*W)=="zinc"
+                      ))
+                        (*W)+='e';
+                      break;
+                    }
+                  }
+                }
+              }
+            }
+            else{
+              W->End=j;
+              return false;
+            }
+          }
+        }
+        W->Type|=TypesStep1b[i];
+        return true;
+      }
+    }
+    return false;
+  }
+  bool Step1c(Word *W){
+    if (W->Length()>2 && tolower((*W)(0))=='y' && IsConsonant((*W)(1))){
+      W->Letters[W->End]='i';
+      return true;
+    }
+    return false;
+  }
+  bool Step2(Word *W, const U32 R1){
+    for (int i=0; i<NUM_SUFFIXES_STEP2; i++){
+      if (W->EndsWith(SuffixesStep2[i][0]) && SuffixInRn(W, R1, SuffixesStep2[i][0])){
+        W->ChangeSuffix(SuffixesStep2[i][0], SuffixesStep2[i][1]);
+        W->Type|=TypesStep2[i];
+        return true;
+      }
+    }
+    if (W->EndsWith("logi") && SuffixInRn(W, R1, "ogi")){
+      W->End--;
+      return true;
+    }
+    else if (W->EndsWith("li")){
+      if (SuffixInRn(W, R1, "li") && IsLiEnding((*W)(2))){
+        W->End-=2;
+        W->Type|=English::AdverbOfManner;
+        return true;
+      }
+      else if (W->Length()>3){
+        switch((*W)(2)){
+            case 'b': {
+              W->Letters[W->End]='e';
+              W->Type|=English::AdverbOfManner;
+              return true;              
+            }
+            case 'i': {
+              if (W->Length()>4){
+                W->End-=2;
+                W->Type|=English::AdverbOfManner;
+                return true;
+              }
+              break;
+            }
+            case 'l': {
+              if (W->Length()>5 && ((*W)(3)=='a' || (*W)(3)=='u')){
+                W->End-=2;
+                W->Type|=English::AdverbOfManner;
+                return true;
+              }
+              break;
+            }
+            case 's': {
+              W->End-=2;
+              W->Type|=English::AdverbOfManner;
+              return true;
+            }
+            case 'e': case 'g': case 'm': case 'n': case 'r': case 'w': {
+              if (W->Length()>(U32)(4+((*W)(2)=='r'))){
+                W->End-=2;
+                W->Type|=English::AdverbOfManner;
+                return true;
+              }
+            }
+        }
+      }
+    }
+    return false;
+  }
+  bool Step3(Word *W, const U32 R1, const U32 R2){
+    bool res=false;
+    for (int i=0; i<NUM_SUFFIXES_STEP3; i++){
+      if (W->EndsWith(SuffixesStep3[i][0]) && SuffixInRn(W, R1, SuffixesStep3[i][0])){
+        W->ChangeSuffix(SuffixesStep3[i][0], SuffixesStep3[i][1]);
+        W->Type|=TypesStep3[i];
+        res=true;
+        break;
+      }
+    }
+    if (W->EndsWith("ative") && SuffixInRn(W, R2, "ative")){
+      W->End-=5;
+      W->Type|=English::SuffixIVE;
+      return true;
+    }
+    if (W->Length()>5 && W->EndsWith("less")){
+      W->End-=4;
+      W->Type|=English::AdjectiveWithout;
+      return true;
+    }
+    return res;
+  }
+  bool Step4(Word *W, const U32 R2){
+    bool res=false;
+    for (int i=0; i<NUM_SUFFIXES_STEP4; i++){
+      if (W->EndsWith(SuffixesStep4[i]) && SuffixInRn(W, R2, SuffixesStep4[i])){
+        W->End-=U8(strlen(SuffixesStep4[i])-(i>17));
+        if (!(i==10 /* ent */ && (*W)(0)=='m'))
+          W->Type|=TypesStep4[i];
+        if (i==0 && W->EndsWith("nti")){
+          W->End--;
+          res=true;
+          continue;
+        }
+        return true;
+      }
+    }
+    return res;
+  }
+  bool Step5(Word *W, const U32 R1, const U32 R2){
+    if ((*W)(0)=='e' && (*W)!="here"){
+      if (SuffixInRn(W, R2, "e"))
+        W->End--;
+      else if (SuffixInRn(W, R1, "e")){
+        W->End--;
+        W->End+=!EndsInShortSyllable(W);
+      }
+      else
+        return false;
+      return true;
+    }
+    else if (W->Length()>1 && (*W)(0)=='l' && SuffixInRn(W, R2, "l") && (*W)(1)=='l'){
+      W->End--;
+      return true;
+    }
+    return false;
+  }
+public:
+  inline bool IsVowel(const char c) final {
+    return CharInArray(c, Vowels, NUM_VOWELS);
+  }
+  inline void Hash(Word *W) final {
+    W->Hash[2] = W->Hash[3] = 0xb0a710ad;
+    for (int i=W->Start; i<=W->End; i++) {
+      U8 l = W->Letters[i];
+      W->Hash[2]=W->Hash[2]*263*32 + l;
+      if (IsVowel(l))
+        W->Hash[3]=W->Hash[3]*997*8 + (l/4-22);
+      else if (l>='b' && l<='z')
+        W->Hash[3]=W->Hash[3]*271*32 + (l-97);
+      else
+        W->Hash[3]=W->Hash[3]*11*32 + l;
+    }
+  }
+  bool Stem(Word *W){
+    if (W->Length()<2){
+      Hash(W);
+      return false;
+    }
+    bool res = TrimStartingApostrophe(W);
+    res|=ProcessPrefixes(W);
+    res|=ProcessSuperlatives(W);
+    for (int i=0; i<NUM_EXCEPTIONS1; i++){
+      if ((*W)==Exceptions1[i][0]){
+        if (i<11){
+          size_t len=strlen(Exceptions1[i][1]);
+          memcpy(&W->Letters[W->Start], Exceptions1[i][1], len);
+          W->End=W->Start+U8(len-1);
+        }
+        Hash(W);
+        W->Type|=TypesExceptions1[i];
+        W->Language = Language::English;
+        return (i<11);
+      }
+    }
+
+    // Start of modified Porter2 Stemmer
+    MarkYsAsConsonants(W);
+    U32 R1=GetRegion1(W), R2=GetRegion(W,R1);
+    res|=Step0(W);
+    res|=Step1a(W);
+    for (int i=0; i<NUM_EXCEPTIONS2; i++){
+      if ((*W)==Exceptions2[i]){
+        Hash(W);
+        W->Type|=TypesExceptions2[i];
+        W->Language = Language::English;
+        return res;
+      }
+    }
+    res|=Step1b(W, R1);
+    res|=Step1c(W);
+    res|=Step2(W, R1);
+    res|=Step3(W, R1, R2);
+    res|=Step4(W, R2);
+    res|=Step5(W, R1, R2);
+
+    for (U8 i=W->Start; i<=W->End; i++){
+      if (W->Letters[i]=='Y')
+        W->Letters[i]='y';
+    }
+    if (!W->Type || W->Type==English::Plural) {
+      if (W->MatchesAny(MaleWords, NUM_MALE_WORDS))
+        res = true, W->Type|=English::Male;
+      else if (W->MatchesAny(FemaleWords, NUM_FEMALE_WORDS))
+        res = true, W->Type|=English::Female;
+    }
+    if (!res)
+      res=W->MatchesAny(CommonWords, NUM_COMMON_WORDS);
+    Hash(W);
+    if (res)
+      W->Language = Language::English;
+    return res;
+  }
+};
+
+
+class FrenchStemmer: public Stemmer {
+private:
+  static const int NUM_VOWELS = 17;
+  const char Vowels[NUM_VOWELS]={'a','e','i','o','u','y','\xE2','\xE0','\xEB','\xE9','\xEA','\xE8','\xEF','\xEE','\xF4','\xFB','\xF9'};
+  static const int NUM_COMMON_WORDS = 10;
+  const char *CommonWords[NUM_COMMON_WORDS]={"de","la","le","et","en","un","une","du","que","pas"};
+  static const int NUM_EXCEPTIONS = 3;
+  const char *(Exceptions[NUM_EXCEPTIONS])[2]={
+    {"monument", "monument"},
+    {"yeux", "oeil"},
+    {"travaux", "travail"},
+  };
+  const U32 TypesExceptions[NUM_EXCEPTIONS]={
+    French::Noun,
+    French::Noun|French::Plural,
+    French::Noun|French::Plural
+  };
+  static const int NUM_SUFFIXES_STEP1 = 39;
+  const char *SuffixesStep1[NUM_SUFFIXES_STEP1]={
+    "ance","iqUe","isme","able","iste","eux","ances","iqUes","ismes","ables","istes", //11
+    "atrice","ateur","ation","atrices","ateurs","ations", //6
+    "logie","logies", //2
+    "usion","ution","usions","utions", //4
+    "ence","ences", //2
+    "issement","issements", //2
+    "ement","ements", //2
+    "it\xE9","it\xE9s", //2
+    "if","ive","ifs","ives", //4
+    "euse","euses", //2
+    "ment","ments" //2
+  };
+  static const int NUM_SUFFIXES_STEP2a = 35;
+  const char *SuffixesStep2a[NUM_SUFFIXES_STEP2a]={
+    "issaIent", "issantes", "iraIent", "issante",
+    "issants", "issions", "irions", "issais",
+    "issait", "issant", "issent", "issiez", "issons",
+    "irais", "irait", "irent", "iriez", "irons",
+    "iront", "isses", "issez", "\xEEmes",
+    "\xEEtes", "irai", "iras", "irez", "isse",
+    "ies", "ira", "\xEEt", "ie", "ir", "is",
+    "it", "i"
+  };
+  static const int NUM_SUFFIXES_STEP2b = 38;
+  const char *SuffixesStep2b[NUM_SUFFIXES_STEP2b]={
+    "eraIent", "assions", "erions", "assent",
+    "assiez", "\xE8rent", "erais", "erait",
+    "eriez", "erons", "eront", "aIent", "antes",
+    "asses", "ions", "erai", "eras", "erez",
+    "\xE2mes", "\xE2tes", "ante", "ants",
+    "asse", "\xE9""es", "era", "iez", "ais",
+    "ait", "ant", "\xE9""e", "\xE9s", "er",
+    "ez", "\xE2t", "ai", "as", "\xE9", "a"
+  };
+  static const int NUM_SET_STEP4 = 6;
+  const char SetStep4[NUM_SET_STEP4]={'a','i','o','u','\xE8','s'};
+  static const int NUM_SUFFIXES_STEP4 = 7;
+  const char *SuffixesStep4[NUM_SUFFIXES_STEP4]={"i\xE8re","I\xE8re","ion","ier","Ier","e","\xEB"};
+  static const int NUM_SUFFIXES_STEP5 = 5;
+  const char *SuffixesStep5[NUM_SUFFIXES_STEP5]={"enn","onn","ett","ell","eill"};
+  inline bool IsConsonant(const char c){
+    return !IsVowel(c);
+  }
+    void ConvertUTF8(Word *W) {
+    for (int i=W->Start; i<W->End; i++) {
+      U8 c = W->Letters[i+1]+((W->Letters[i+1]<0xA0)?0x60:0x40);
+      if (W->Letters[i]==0xC3 && (IsVowel(c) || (W->Letters[i+1]&0xDF)==0x87)) {
+        W->Letters[i] = c;
+        if (i+1<W->End)
+          memmove(&W->Letters[i+1], &W->Letters[i+2], W->End-i-1);
+        W->End--;
+      }
+    }
+  }
+  void MarkVowelsAsConsonants(Word *W){
+    for (int i=W->Start; i<=W->End; i++){
+      switch (W->Letters[i]) {
+        case 'i': case 'u': {
+          if (i>W->Start && i<W->End && (IsVowel(W->Letters[i-1]) || (W->Letters[i-1]=='q' && W->Letters[i]=='u')) && IsVowel(W->Letters[i+1]))
+            W->Letters[i] = toupper(W->Letters[i]);
+          break;
+        }
+        case 'y': {
+          if ((i>W->Start && IsVowel(W->Letters[i-1])) || (i<W->End && IsVowel(W->Letters[i+1])))
+            W->Letters[i] = toupper(W->Letters[i]);
+        }
+      }
+    }
+  }
+  U32 GetRV(Word *W) {
+    U32 len = W->Length(), res = W->Start+len;
+    if (len>=3 && ((IsVowel(W->Letters[W->Start]) && IsVowel(W->Letters[W->Start+1])) || W->StartsWith("par") || W->StartsWith("col") || W->StartsWith("tap") ))
+      return W->Start+3;
+    else {
+      for (int i=W->Start+1;i<=W->End;i++) {
+        if (IsVowel(W->Letters[i]))
+          return i+1;
+      }
+    }
+    return res;
+  }
+     bool Step1(Word *W, const U32 RV, const U32 R1, const U32 R2, bool *ForceStep2a) {
+    int i = 0;
+    for (; i<11; i++) {
+      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R2, SuffixesStep1[i])) {
+        W->End-=U8(strlen(SuffixesStep1[i]));
+        if (i==3 /*able*/)
+          W->Type|=French::Adjective;
+        return true;
+      }
+    }
+    for (; i<17; i++) {
+      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R2, SuffixesStep1[i])) {
+        W->End-=U8(strlen(SuffixesStep1[i]));
+        if (W->EndsWith("ic"))
+          W->ChangeSuffix("c", "qU");
+        return true;
+      }
+    }
+    for (; i<25;i++) {
+      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R2, SuffixesStep1[i])) {
+        W->End-=U8(strlen(SuffixesStep1[i]))-1-(i<19)*2;
+        if (i>22) {
+          W->End+=2;
+          W->Letters[W->End]='t';
+        }
+        return true;
+      }
+    }
+    for (; i<27; i++) {
+      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R1, SuffixesStep1[i]) && IsConsonant((*W)((U8)strlen(SuffixesStep1[i])))) {
+        W->End-=U8(strlen(SuffixesStep1[i]));
+        return true;
+      }
+    }
+    for (; i<29; i++) {
+      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, RV, SuffixesStep1[i])) {
+        W->End-=U8(strlen(SuffixesStep1[i]));
+        if (W->EndsWith("iv") && SuffixInRn(W, R2, "iv")) {
+          W->End-=2;
+          if (W->EndsWith("at") && SuffixInRn(W, R2, "at"))
+            W->End-=2;
+        }
+        else if (W->EndsWith("eus")) {
+          if (SuffixInRn(W, R2, "eus"))
+            W->End-=3;
+          else if (SuffixInRn(W, R1, "eus"))
+            W->Letters[W->End]='x';
+        }
+        else if ((W->EndsWith("abl") && SuffixInRn(W, R2, "abl")) || (W->EndsWith("iqU") && SuffixInRn(W, R2, "iqU")))
+          W->End-=3;
+        else if ((W->EndsWith("i\xE8r") && SuffixInRn(W, RV, "i\xE8r")) || (W->EndsWith("I\xE8r") && SuffixInRn(W, RV, "I\xE8r"))) {
+          W->End-=2;
+          W->Letters[W->End]='i';
+        }
+        return true;
+      }
+    }
+    for (; i<31; i++) {
+      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R2, SuffixesStep1[i])) {
+        W->End-=U8(strlen(SuffixesStep1[i]));
+        if (W->EndsWith("abil")) {
+          if (SuffixInRn(W, R2, "abil"))
+            W->End-=4;
+          else
+            W->End--, W->Letters[W->End]='l';
+        }
+        else if (W->EndsWith("ic")) {
+          if (SuffixInRn(W, R2, "ic"))
+            W->End-=2;
+          else
+            W->ChangeSuffix("c", "qU");
+        }
+        else if (W->EndsWith("iv") && SuffixInRn(W, R2, "iv"))
+          W->End-=2;
+        return true;
+      }
+    }
+    for (; i<35; i++) {
+      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R2, SuffixesStep1[i])) {
+        W->End-=U8(strlen(SuffixesStep1[i]));
+        if (W->EndsWith("at") && SuffixInRn(W, R2, "at")) {
+          W->End-=2;
+          if (W->EndsWith("ic")) {
+            if (SuffixInRn(W, R2, "ic"))
+              W->End-=2;
+            else
+              W->ChangeSuffix("c", "qU");
+          }
+        }
+        return true;
+      }
+    }
+    for (; i<37; i++) {
+      if (W->EndsWith(SuffixesStep1[i])) {
+        if (SuffixInRn(W, R2, SuffixesStep1[i])) {
+          W->End-=U8(strlen(SuffixesStep1[i]));
+          return true;
+        }
+        else if (SuffixInRn(W, R1, SuffixesStep1[i])) {
+          W->ChangeSuffix(SuffixesStep1[i], "eux");
+          return true;
+        }
+      }
+    }
+    for (; i<NUM_SUFFIXES_STEP1; i++) {
+      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, RV+1, SuffixesStep1[i]) && IsVowel((*W)((U8)strlen(SuffixesStep1[i])))) {
+        W->End-=U8(strlen(SuffixesStep1[i]));
+        (*ForceStep2a) = true;
+        return true;
+      }
+    }
+    if (W->EndsWith("eaux") || (*W)=="eaux") {
+      W->End--;
+      W->Type|=French::Plural;
+      return true;
+    }
+    else if (W->EndsWith("aux") && SuffixInRn(W, R1, "aux")) {
+      W->End--, W->Letters[W->End] = 'l';
+      W->Type|=French::Plural;
+      return true;
+    }
+    else if (W->EndsWith("amment") && SuffixInRn(W, RV, "amment")) {
+      W->ChangeSuffix("amment", "ant");
+      (*ForceStep2a) = true;
+      return true;
+    }
+    else if (W->EndsWith("emment") && SuffixInRn(W, RV, "emment")) {
+      W->ChangeSuffix("emment", "ent");
+      (*ForceStep2a) = true;
+      return true;
+    }
+    return false;
+  }
+  bool Step2a(Word *W, const U32 RV) {
+    for (int i=0; i<NUM_SUFFIXES_STEP2a; i++) {
+      if (W->EndsWith(SuffixesStep2a[i]) && SuffixInRn(W, RV+1, SuffixesStep2a[i]) && IsConsonant((*W)((U8)strlen(SuffixesStep2a[i])))) {
+        W->End-=U8(strlen(SuffixesStep2a[i]));
+        if (i==31 /*ir*/)
+          W->Type|=French::Verb;
+        return true;
+      }
+    }
+    return false;
+  }
+  bool Step2b(Word *W, const U32 RV, const U32 R2) {
+    for (int i=0; i<NUM_SUFFIXES_STEP2b; i++) {
+      if (W->EndsWith(SuffixesStep2b[i]) && SuffixInRn(W, RV, SuffixesStep2b[i])) {
+        switch (SuffixesStep2b[i][0]) {
+          case 'a': case '\xE2': {
+            W->End-=U8(strlen(SuffixesStep2b[i]));
+            if (W->EndsWith("e") && SuffixInRn(W, RV, "e"))
+              W->End--;
+            return true;
+          }
+          default: {
+            if (i!=14 || SuffixInRn(W, R2, SuffixesStep2b[i])) {
+              W->End-=U8(strlen(SuffixesStep2b[i]));
+              return true;
+            }
+          }
+        }        
+      }
+    }
+    return false;
+  }
+  void Step3(Word *W) {
+    char *final = (char *)&W->Letters[W->End];
+    if ((*final)=='Y')
+      (*final) = 'i';
+    else if ((*final)=='\xE7')
+      (*final) = 'c';
+  }
+  bool Step4(Word *W, const U32 RV, const U32 R2) {
+    bool res = false;
+    if (W->Length()>=2 && W->Letters[W->End]=='s' && !CharInArray((*W)(1), SetStep4, NUM_SET_STEP4)) {
+      W->End--;
+      res = true;
+    }
+    for (int i=0; i<NUM_SUFFIXES_STEP4; i++) {
+      if (W->EndsWith(SuffixesStep4[i]) && SuffixInRn(W, RV, SuffixesStep4[i])) {
+        switch (i) {
+          case 2: { //ion
+            char prec = (*W)(3);
+            if (SuffixInRn(W, R2, SuffixesStep4[i]) && SuffixInRn(W, RV+1, SuffixesStep4[i]) && (prec=='s' || prec=='t')) {
+              W->End-=3;
+              return true;
+            }
+            break;
+          }
+          case 5: { //e
+            W->End--;
+            return true;
+          }
+          case 6: { //\xEB
+            if (W->EndsWith("gu\xEB")) {
+              W->End--;
+              return true;
+            }
+            break;
+          }
+          default: {
+            W->ChangeSuffix(SuffixesStep4[i], "i");
+            return true;
+          }
+        }
+      }
+    }
+    return res;
+  }
+  bool Step5(Word *W) {
+    for (int i=0; i<NUM_SUFFIXES_STEP5; i++) {
+      if (W->EndsWith(SuffixesStep5[i])) {
+        W->End--;
+        return true;
+      }
+    }
+    return false;
+  }
+  bool Step6(Word *W) {
+    for (int i=W->End; i>=W->Start; i--) {
+      if (IsVowel(W->Letters[i])) {
+        if (i<W->End && (W->Letters[i]&0xFE)==0xE8) {
+          W->Letters[i] = 'e';
+          return true;
+        }
+        return false;
+      }
+    }
+    return false;
+  }
+public:
+  inline bool IsVowel(const char c) final {
+    return CharInArray(c, Vowels, NUM_VOWELS);
+  }
+  inline void Hash(Word *W) final {
+    W->Hash[2] = W->Hash[3] = ~0xeff1cace;
+    for (int i=W->Start; i<=W->End; i++) {
+      U8 l = W->Letters[i];
+      W->Hash[2]=W->Hash[2]*251*32 + l;
+      if (IsVowel(l))
+        W->Hash[3]=W->Hash[3]*997*16 + l;
+      else if (l>='b' && l<='z')
+        W->Hash[3]=W->Hash[3]*271*32 + (l-97);
+      else
+        W->Hash[3]=W->Hash[3]*11*32 + l;
+    }
+  }
+  bool Stem(Word *W) {
+    ConvertUTF8(W);
+    if (W->Length()<2) {
+      Hash(W);
+      return false;
+    }
+    for (int i=0; i<NUM_EXCEPTIONS; i++) {
+      if ((*W)==Exceptions[i][0]) {
+        size_t len=strlen(Exceptions[i][1]);
+        memcpy(&W->Letters[W->Start], Exceptions[i][1], len);
+        W->End=W->Start+U8(len-1);
+        Hash(W);
+        W->Type|=TypesExceptions[i];
+        W->Language = Language::French;
+        return true;
+      }
+    }
+    MarkVowelsAsConsonants(W);
+    U32 RV=GetRV(W), R1=GetRegion(W, 0), R2=GetRegion(W, R1);
+    bool DoNextStep=false, res=Step1(W, RV, R1, R2, &DoNextStep);
+    DoNextStep|=!res;
+    if (DoNextStep) {
+      DoNextStep = !Step2a(W, RV);
+      res|=!DoNextStep;
+      if (DoNextStep)
+        res|=Step2b(W, RV, R2);
+    }
+    if (res)
+      Step3(W);
+    else
+      res|=Step4(W, RV, R2);
+    res|=Step5(W);
+    res|=Step6(W);
+    for (int i=W->Start; i<=W->End; i++)
+      W->Letters[i] = tolower(W->Letters[i]);
+    if (!res)
+      res=W->MatchesAny(CommonWords, NUM_COMMON_WORDS);
+    Hash(W);
+    if (res)
+      W->Language = Language::French;
+    return res;
+  }
+};
+
+class GermanStemmer : public Stemmer {
+private:
+  static const int NUM_VOWELS = 9;
+  const char Vowels[NUM_VOWELS]={'a','e','i','o','u','y','\xE4','\xF6','\xFC'};
+  static const int NUM_COMMON_WORDS = 10;
+  const char *CommonWords[NUM_COMMON_WORDS]={"der","die","das","und","sie","ich","mit","sich","auf","nicht"};
+  static const int NUM_ENDINGS = 10;
+  const char Endings[NUM_ENDINGS]={'b','d','f','g','h','k','l','m','n','t'}; //plus 'r' for words ending in 's'
+  static const int NUM_SUFFIXES_STEP1 = 6;
+  const char *SuffixesStep1[NUM_SUFFIXES_STEP1]={"em","ern","er","e","en","es"};
+  static const int NUM_SUFFIXES_STEP2 = 3;
+  const char *SuffixesStep2[NUM_SUFFIXES_STEP2]={"en","er","est"};
+  static const int NUM_SUFFIXES_STEP3 = 7;
+  const char *SuffixesStep3[NUM_SUFFIXES_STEP3]={"end","ung","ik","ig","isch","lich","heit"};
+  void ConvertUTF8(Word *W) {
+    for (int i=W->Start; i<W->End; i++) {
+      U8 c = W->Letters[i+1]+((W->Letters[i+1]<0x9F)?0x60:0x40);
+      if (W->Letters[i]==0xC3 && (IsVowel(c) || c==0xDF)) {
+        W->Letters[i] = c;
+        if (i+1<W->End)
+          memmove(&W->Letters[i+1], &W->Letters[i+2], W->End-i-1);
+        W->End--;
+      }
+    }
+  }
+  void ReplaceSharpS(Word *W) {
+    for (int i=W->Start; i<=W->End; i++) {
+      if (W->Letters[i]==0xDF) {
+        W->Letters[i]='s';
+        if (i+1<MAX_WORD_SIZE) {
+          memmove(&W->Letters[i+2], &W->Letters[i+1], MAX_WORD_SIZE-i-2);
+          W->Letters[i+1]='s';
+          W->End+=(W->End<MAX_WORD_SIZE-1);
+        }
+      }
+    }
+  }    
+  void MarkVowelsAsConsonants(Word *W) {
+    for (int i=W->Start+1; i<W->End; i++) {
+      U8 c = W->Letters[i];
+      if ((c=='u' || c=='y') && IsVowel(W->Letters[i-1]) && IsVowel(W->Letters[i+1]))
+        W->Letters[i] = toupper(c);
+    }
+  }
+  inline bool IsValidEnding(const char c, const bool IncludeR = false) {
+    return CharInArray(c, Endings, NUM_ENDINGS) || (IncludeR && c=='r');
+  }
+  bool Step1(Word *W, const U32 R1) {
+    int i = 0;
+    for (; i<3; i++) {
+      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R1, SuffixesStep1[i])) {
+        W->End-=U8(strlen(SuffixesStep1[i]));
+        return true;
+      }
+    }
+    for (; i<NUM_SUFFIXES_STEP1; i++) {
+      if (W->EndsWith(SuffixesStep1[i]) && SuffixInRn(W, R1, SuffixesStep1[i])) {
+        W->End-=U8(strlen(SuffixesStep1[i]));
+        W->End-=U8(W->EndsWith("niss"));
+        return true;
+      }
+    }
+    if (W->EndsWith("s") && SuffixInRn(W, R1, "s") && IsValidEnding((*W)(1), true)) {
+      W->End--;
+      return true;
+    }
+    return false;
+  }
+  bool Step2(Word *W, const U32 R1) {
+    for (int i=0; i<NUM_SUFFIXES_STEP2; i++) {
+      if (W->EndsWith(SuffixesStep2[i]) && SuffixInRn(W, R1, SuffixesStep2[i])) {
+        W->End-=U8(strlen(SuffixesStep2[i]));
+        return true;
+      }
+    }
+    if (W->EndsWith("st") && SuffixInRn(W, R1, "st") && W->Length()>5 && IsValidEnding((*W)(2))) {
+      W->End-=2;
+      return true;
+    }
+    return false;
+  }
+  bool Step3(Word *W, const U32 R1, const U32 R2) {
+    int i = 0;
+    for (; i<2; i++) {
+      if (W->EndsWith(SuffixesStep3[i]) && SuffixInRn(W, R2, SuffixesStep3[i])) {
+        W->End-=U8(strlen(SuffixesStep3[i]));
+        if (W->EndsWith("ig") && (*W)(2)!='e' && SuffixInRn(W, R2, "ig"))
+          W->End-=2;
+        if (i)
+          W->Type|=German::Noun;
+        return true;
+      }
+    }
+    for (; i<5; i++) {
+      if (W->EndsWith(SuffixesStep3[i]) && SuffixInRn(W, R2, SuffixesStep3[i]) && (*W)((U8)strlen(SuffixesStep3[i]))!='e') {
+        W->End-=U8(strlen(SuffixesStep3[i]));
+        if (i>2)
+          W->Type|=German::Adjective;
+        return true;
+      }
+    }
+    for (; i<NUM_SUFFIXES_STEP3; i++) {
+      if (W->EndsWith(SuffixesStep3[i]) && SuffixInRn(W, R2, SuffixesStep3[i])) {
+        W->End-=U8(strlen(SuffixesStep3[i]));
+        if ((W->EndsWith("er") || W->EndsWith("en")) && SuffixInRn(W, R1, "e?"))
+          W->End-=2;
+        if (i>5)
+          W->Type|=German::Noun|German::Female;
+        return true;
+      }
+    }
+    if (W->EndsWith("keit") && SuffixInRn(W, R2, "keit")) {
+      W->End-=4;
+      if (W->EndsWith("lich") && SuffixInRn(W, R2, "lich"))
+        W->End-=4;
+      else if (W->EndsWith("ig") && SuffixInRn(W, R2, "ig"))
+        W->End-=2;
+      W->Type|=German::Noun|German::Female;
+      return true;
+    }
+    return false;
+  }
+public:
+  inline bool IsVowel(const char c) final {
+    return CharInArray(c, Vowels, NUM_VOWELS);
+  }
+  inline void Hash(Word *W) final {
+    W->Hash[2] = W->Hash[3] = ~0xbea7ab1e;
+    for (int i=W->Start; i<=W->End; i++) {
+      U8 l = W->Letters[i];
+      W->Hash[2]=W->Hash[2]*263*32 + l;
+      if (IsVowel(l))
+        W->Hash[3]=W->Hash[3]*997*16 + l;
+      else if (l>='b' && l<='z')
+        W->Hash[3]=W->Hash[3]*251*32 + (l-97);
+      else
+        W->Hash[3]=W->Hash[3]*11*32 + l;
+    }
+  }
+  bool Stem(Word *W) {
+    ConvertUTF8(W);
+    if (W->Length()<2) {
+      Hash(W);
+      return false;
+    }
+    ReplaceSharpS(W);
+    MarkVowelsAsConsonants(W);
+    U32 R1=GetRegion(W, 0), R2=GetRegion(W, R1);
+    R1 = min(3, R1);
+    bool res = Step1(W, R1);
+    res|=Step2(W, R1);
+    res|=Step3(W, R1, R2);
+    for (int i=W->Start; i<=W->End; i++) {
+      switch (W->Letters[i]) {
+        case 0xE4: { W->Letters[i] = 'a'; break; }
+        case 0xF6: case 0xFC: { W->Letters[i]-=0x87; break; }
+        default: W->Letters[i] = tolower(W->Letters[i]);
+      }
+    }
+    if (!res)
+      res=W->MatchesAny(CommonWords, NUM_COMMON_WORDS);
+    Hash(W);
+    if (res)
+      W->Language = Language::German;
+    return res;
+  }
+};
